Refactoring changes push 1
Renaming, cleaning up, and moving things around in preparation for project handoff. More pushes to come.
