--- .VSCodeCounter/2022-03-08_16-56-26/diff-details.md
@@ -1,15 +1,0 @@
-# Diff Details
-
-Date : 2022-03-08 16:56:26
-
-Directory c:\Users\ryan\Desktop\Ryan's Folder\Github\examples
-
-Total : 0 files,  0 codes, 0 comments, 0 blanks, all 0 lines
-
-[summary](results.md) / [details](details.md) / [diff summary](diff.md) / diff details
-
-## Files
-| filename | language | code | comment | blank | total |
-| :--- | :--- | ---: | ---: | ---: | ---: |
-
-[summary](results.md) / [details](details.md) / [diff summary](diff.md) / diff details
