--- .VSCodeCounter/2022-03-08_16-56-26/diff.csv
@@ -1,2 +1,0 @@
-"filename", "language", "", "comment", "blank", "total"
-"Total", "-", , 0, 0, 0
