--- .VSCodeCounter/2022-03-08_16-56-26/diff.md
@@ -1,19 +1,0 @@
-# Diff Summary
-
-Date : 2022-03-08 16:56:26
-
-Directory c:\Users\ryan\Desktop\Ryan's Folder\Github\examples
-
-Total : 0 files,  0 codes, 0 comments, 0 blanks, all 0 lines
-
-[summary](results.md) / [details](details.md) / diff summary / [diff details](diff-details.md)
-
-## Languages
-| language | files | code | comment | blank | total |
-| :--- | ---: | ---: | ---: | ---: | ---: |
-
-## Directories
-| path | files | code | comment | blank | total |
-| :--- | ---: | ---: | ---: | ---: | ---: |
-
-[summary](results.md) / [details](details.md) / diff summary / [diff details](diff-details.md)
