--- .VSCodeCounter/2022-03-08_16-56-26/diff.txt
@@ -1,22 +1,0 @@
-Date : 2022-03-08 16:56:26
-Directory : c:\Users\ryan\Desktop\Ryan's Folder\Github\examples
-Total : 0 files,  0 codes, 0 comments, 0 blanks, all 0 lines
-
-Languages
-+----------+------------+------------+------------+------------+------------+
-| language | files      | code       | comment    | blank      | total      |
-+----------+------------+------------+------------+------------+------------+
-+----------+------------+------------+------------+------------+------------+
-
-Directories
-+------+------------+------------+------------+------------+------------+
-| path | files      | code       | comment    | blank      | total      |
-+------+------------+------------+------------+------------+------------+
-+------+------------+------------+------------+------------+------------+
-
-Files
-+----------+----------+------------+------------+------------+------------+
-| filename | language | code       | comment    | blank      | total      |
-+----------+----------+------------+------------+------------+------------+
-| Total    |          |          0 |          0 |          0 |          0 |
-+----------+----------+------------+------------+------------+------------+
