--- .VSCodeCounter/2022-03-08_16-56-26/results.md
@@ -1,46 +1,0 @@
-# Summary
-
-Date : 2022-03-08 16:56:26
-
-Directory c:\Users\ryan\Desktop\Ryan's Folder\Github\examples
-
-Total : 159 files,  53917 codes, 1381 comments, 2686 blanks, all 57984 lines
-
-summary / [details](details.md) / [diff summary](diff.md) / [diff details](diff-details.md)
-
-## Languages
-| language | files | code | comment | blank | total |
-| :--- | ---: | ---: | ---: | ---: | ---: |
-| JavaScript | 139 | 35,061 | 1,363 | 2,593 | 39,017 |
-| JSON | 8 | 18,588 | 0 | 3 | 18,591 |
-| Markdown | 3 | 127 | 0 | 53 | 180 |
-| CSS | 2 | 44 | 0 | 4 | 48 |
-| YAML | 2 | 32 | 5 | 6 | 43 |
-| Docker | 1 | 29 | 9 | 8 | 46 |
-| HTML | 1 | 22 | 1 | 13 | 36 |
-| Shell Script | 3 | 14 | 3 | 6 | 23 |
-
-## Directories
-| path | files | code | comment | blank | total |
-| :--- | ---: | ---: | ---: | ---: | ---: |
-| . | 159 | 53,917 | 1,381 | 2,686 | 57,984 |
-| codebuild | 1 | 17 | 5 | 5 | 27 |
-| config | 1 | 9 | 4 | 0 | 13 |
-| public | 3 | 2,802 | 0 | 0 | 2,802 |
-| public\arrivals | 2 | 2,170 | 0 | 0 | 2,170 |
-| public\locations | 1 | 632 | 0 | 0 | 632 |
-| scripts | 3 | 14 | 3 | 6 | 23 |
-| src | 141 | 37,791 | 1,320 | 2,578 | 41,689 |
-| src\arrivals | 1 | 2,085 | 0 | 0 | 2,085 |
-| src\assets | 19 | 28,849 | 420 | 1,000 | 30,269 |
-| src\assets\arrivals | 7 | 7,853 | 420 | 4 | 8,277 |
-| src\assets\locations | 4 | 1,538 | 0 | 0 | 1,538 |
-| src\assets\objs | 6 | 19,409 | 0 | 992 | 20,401 |
-| src\behavior | 63 | 2,046 | 375 | 636 | 3,057 |
-| src\behavior\responsibility | 35 | 1,006 | 158 | 292 | 1,456 |
-| src\locations | 1 | 632 | 0 | 0 | 632 |
-| src\people | 36 | 3,329 | 385 | 739 | 4,453 |
-| src\support | 12 | 600 | 56 | 148 | 804 |
-| test | 2 | 31 | 39 | 20 | 90 |
-
-summary / [details](details.md) / [diff summary](diff.md) / [diff details](diff-details.md)
