--- src/assets/arrivals/arrivalHospitalThesis.js
@@ -1,1015 +1,0 @@
-export default [
-  {
-    "name": "patient",
-    "arrivalTick": 0,
-    "arrivalLocation": "Main Entrance",
-    "age": 34,
-    "severity": "ESI3",
-    "patientName": "Daniel",
-    "gender": "male",
-    "id": 0
-  },
-  // THESE ARE THE 6 AM NURSES
-  {
-    "name": "Nurse",
-    "type": "Triage Nurse",
-    "arrivalLocation": "Main Entrance",
-    "arrivalTick": 10,
-    "id": 1
-  },
-  {
-    "name": "Nurse",
-    "type": "Triage Nurse",
-    "arrivalLocation": "Main Entrance",
-    "arrivalTick": 10,
-    "id": 2
-  },
-  {
-    "name": "Tech",
-    "type": "Tech",
-    "arrivalLocation": "Main Entrance",
-    "arrivalTick": 10,
-    "id": 3
-  },
-  {
-    "name": "Tech",
-    "type": "Tech",
-    "arrivalLocation": "Main Entrance",
-    "arrivalTick": 10,
-    "id": 4
-  },
-  {
-    "name": "Nurse",
-    "type": "Nurse",
-    "arrivalLocation": "Main Entrance",
-    "arrivalTick": 10,
-    "id": 5
-  },
-  {
-    "name": "Nurse",
-    "type": "Nurse",
-    "arrivalLocation": "Main Entrance",
-    "arrivalTick": 10,
-    "id": 6
-  },
-  {
-    "name": "Nurse",
-    "type": "Greeter Nurse",
-    "arrivalLocation": "Main Entrance",
-    "arrivalTick": 10,
-    "id": 7
-  },
-  {
-    "name": "Tech",
-    "type": "CT",
-    "arrivalLocation": "Main Entrance",
-    "arrivalTick": 10,
-    "id": 8
-  },
-  {
-    "name": "Tech",
-    "type": "CT",
-    "arrivalLocation": "Main Entrance",
-    "arrivalTick": 10,
-    "id": 9
-  },
-  {
-    "name": "Tech",
-    "type": "Janitorial",
-    "arrivalLocation": "Main Entrance",
-    "arrivalTick": 10,
-    "id": 10
-  },
-  {
-    "name": "Tech",
-    "type": "Radiology",
-    "arrivalLocation": "Main Entrance",
-    "arrivalTick": 10,
-    "id": 11
-  },
-  {
-    "name": "Attending",
-    "type": "Attending",
-    "arrivalLocation": "Main Entrance",
-    "arrivalTick": 20,
-    "id": 12
-  },
-  {
-    "name": "Resident",
-    "type": "Resident",
-    "arrivalLocation": "Main Entrance",
-    "arrivalTick": 20,
-    "id": 13
-  },
-  {
-    "name": "Resident",
-    "type": "Resident",
-    "arrivalLocation": "Main Entrance",
-    "arrivalTick": 20,
-    "id": 14
-  },
-  {
-    "name": "Pharmacist",
-    "type": "Pharmacist",
-    "arrivalLocation": "Main Entrance",
-    "arrivalTick": 30,
-    "id": 15
-  },
-  {
-    "name": "Tech",
-    "type": "XRay",
-    "arrivalLocation": "Main Entrance",
-    "arrivalTick": 40,
-    "id": 16
-  },
-  {
-    "name": "Tech",
-    "type": "XRay",
-    "arrivalLocation": "Main Entrance",
-    "arrivalTick": 40,
-    "id": 17
-  },
-  {
-    "name": "Tech",
-    "type": "Phlebotomist",
-    "arrivalLocation": "Main Entrance",
-    "arrivalTick": 50,
-    "id": 18
-  },
-  {
-    "name": "Tech",
-    "type": "Phlebotomist",
-    "arrivalLocation": "Main Entrance",
-    "arrivalTick": 50,
-    "id": 19
-  },
-  {
-    "name": "Tech",
-    "type": "Phlebotomist",
-    "arrivalLocation": "Main Entrance",
-    "arrivalTick": 50,
-    "id": 20
-  },
-  {
-    "name": "Tech",
-    "type": "Phlebotomist",
-    "arrivalLocation": "Main Entrance",
-    "arrivalTick": 50,
-    "id": 21
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 30057,
-    "arrivalLocation": "Main Entrance",
-    "age": 39,
-    "severity": "ESI3",
-    "patientName": "James",
-    "gender": "male",
-    "id": 22
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 77178,
-    "arrivalLocation": "Main Entrance",
-    "age": 69,
-    // "severity": "ESI3",
-    "severity": "ESI3",
-    "patientName": "Susan",
-    "gender": "female",
-    "id": 23
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 128609,
-    "arrivalLocation": "Main Entrance",
-    "age": 38,
-    "severity": "ESI3",
-    "patientName": "Michael",
-    "gender": "male",
-    "id": 24
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 132392,
-    "arrivalLocation": "Main Entrance",
-    "age": 79,
-    "severity": "ESI3",
-    "patientName": "James",
-    "gender": "male",
-    "id": 25
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 234701,
-    "arrivalLocation": "Main Entrance",
-    "age": 11,
-    "severity": "ESI3",
-    "patientName": "Ethan",
-    "gender": "male",
-    "id": 26
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 324000,
-    "arrivalLocation": "Main Entrance",
-    "age": 21,
-    "severity": "ESI3",
-    "patientName": "Michael",
-    "gender": "male",
-    "id": 27
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 432000,
-    //"arrivalLocation": "Ambulance Entrance",
-    "arrivalLocation": "Main Entrance",
-    "age": 39,
-    //"severity": "ESI1",
-    "severity": "ESI3",
-    "patientName": "David",
-    "gender": "male",
-    "id": 28
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 460000,
-    "arrivalLocation": "Main Entrance",
-    "age": 4,
-    "severity": "ESI2",
-    "patientName": "Sophia",
-    "gender": "female",
-    "id": 29
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 578782,
-    "arrivalLocation": "Main Entrance",
-    "age": 12,
-    "severity": "ESI2",
-    "patientName": "Joshua",
-    "gender": "male",
-    "id": 30
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 648000,
-    "arrivalLocation": "Main Entrance",
-    "age": 2,
-    "severity": "ESI2",
-    "patientName": "Ava",
-    "gender": "female",
-    "id": 31
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 756000,
-    "arrivalLocation": "Main Entrance",
-    "age": 52,
-    "severity": "ESI2",
-    "patientName": "Kimberly",
-    "gender": "female",
-    "id": 32
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 809758,
-    "arrivalLocation": "Main Entrance",
-    "age": 48,
-    "severity": "ESI3",
-    "patientName": "Jennifer",
-    "gender": "female",
-    "id": 33
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 876574,
-    "arrivalLocation": "Main Entrance",
-    "age": 32,
-    "severity": "ESI3",
-    "patientName": "Jessica",
-    "gender": "female",
-    "id": 34
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 888214,
-    "arrivalLocation": "Main Entrance",
-    "age": 23,
-    "severity": "ESI3",
-    "patientName": "Joshua",
-    "gender": "male",
-    "id": 35
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 972000,
-    "arrivalLocation": "Main Entrance",
-    "age": 18,
-    "severity": "ESI2",
-    "patientName": "Jacob",
-    "gender": "male",
-    "id": 36
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1001661,
-    "arrivalLocation": "Main Entrance",
-    "age": 65,
-    "severity": "ESI2",
-    "patientName": "Linda",
-    "gender": "female",
-    "id": 37
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1033598,
-    "arrivalLocation": "Main Entrance",
-    "age": 59,
-    "severity": "ESI3",
-    "patientName": "Donna",
-    "gender": "female",
-    "id": 38
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1045057,
-    "arrivalLocation": "Main Entrance",
-    "age": 43,
-    "severity": "ESI2",
-    "patientName": "Michael",
-    "gender": "male",
-    "id": 39
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1094277,
-    "arrivalLocation": "Main Entrance",
-    "age": 20,
-    "severity": "ESI3",
-    "patientName": "Sarah",
-    "gender": "female",
-    "id": 40
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1120783,
-    "arrivalLocation": "Main Entrance",
-    "age": 44,
-    "severity": "ESI2",
-    "patientName": "David",
-    "gender": "male",
-    "id": 41
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1169077,
-    "arrivalLocation": "Main Entrance",
-    "age": 1,
-    "severity": "ESI3",
-    "patientName": "Oliver",
-    "gender": "male",
-    "id": 42
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1184194,
-    "arrivalLocation": "Main Entrance",
-    "age": 22,
-    "severity": "ESI3",
-    "patientName": "Christopher",
-    "gender": "male",
-    "id": 43
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1206684,
-    "arrivalLocation": "Main Entrance",
-    "age": 9,
-    "severity": "ESI2",
-    "patientName": "Olivia",
-    "gender": "female",
-    "id": 44
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1236019,
-    //"arrivalLocation": "Ambulance Entrance",
-    "arrivalLocation": "Main Entrance",
-    "age": 2,
-    //"severity": "ESI1",
-    "severity": "ESI3",
-    "patientName": "Isabella",
-    "gender": "female",
-    "id": 45
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1264071,
-    "arrivalLocation": "Main Entrance",
-    "age": 23,
-    "severity": "ESI3",
-    "patientName": "Michael",
-    "gender": "male",
-    "id": 46
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1276884,
-    "arrivalLocation": "Main Entrance",
-    "age": 38,
-    "severity": "ESI3",
-    "patientName": "Matthew",
-    "gender": "male",
-    "id": 47
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1311043,
-    "arrivalLocation": "Main Entrance",
-    "age": 66,
-    "severity": "ESI3",
-    "patientName": "Robert",
-    "gender": "male",
-    "id": 48
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1327346,
-    "arrivalLocation": "Main Entrance",
-    "age": 84,
-    "severity": "ESI2",
-    "patientName": "William",
-    "gender": "male",
-    "id": 49
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1368891,
-    "arrivalLocation": "Main Entrance",
-    "age": 9,
-    "severity": "ESI3",
-    "patientName": "Jayden",
-    "gender": "male",
-    "id": 50
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1394895,
-    "arrivalLocation": "Main Entrance",
-    "age": 40,
-    "severity": "ESI3",
-    "patientName": "Jennifer",
-    "gender": "female",
-    "id": 51
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1416461,
-    "arrivalLocation": "Main Entrance",
-    "age": 66,
-    "severity": "ESI3",
-    "patientName": "Robert",
-    "gender": "male",
-    "id": 52
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1457151,
-    "arrivalLocation": "Main Entrance",
-    "age": 67,
-    "severity": "ESI2",
-    "patientName": "John",
-    "gender": "male",
-    "id": 53
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1475659,
-    "arrivalLocation": "Main Entrance",
-    "age": 75,
-    "severity": "ESI3",
-    "patientName": "James",
-    "gender": "male",
-    "id": 54
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1506525,
-    "arrivalLocation": "Main Entrance",
-    "age": 16,
-    "severity": "ESI2",
-    "patientName": "Hannah",
-    "gender": "female",
-    "id": 55
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1538080,
-    "arrivalLocation": "Main Entrance",
-    "age": 72,
-    "severity": "ESI3",
-    "patientName": "Barbara",
-    "gender": "female",
-    "id": 56
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1539842,
-    "arrivalLocation": "Main Entrance",
-    "age": 6,
-    "severity": "ESI2",
-    "patientName": "Isabella",
-    "gender": "female",
-    "id": 57
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1577255,
-    "arrivalLocation": "Main Entrance",
-    "age": 50,
-    "severity": "ESI2",
-    "patientName": "Melissa",
-    "gender": "female",
-    "id": 58
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1620000,
-    "arrivalLocation": "Main Entrance",
-    "age": 23,
-    "severity": "ESI3",
-    "patientName": "Matthew",
-    "gender": "male",
-    "id": 59
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1671995,
-    "arrivalLocation": "Main Entrance",
-    "age": 14,
-    "severity": "ESI3",
-    "patientName": "Madison",
-    "gender": "female",
-    "id": 60
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1690977,
-    "arrivalLocation": "Main Entrance",
-    "age": 12,
-    "severity": "ESI3",
-    "patientName": "Ethan",
-    "gender": "male",
-    "id": 61
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1713298,
-    "arrivalLocation": "Main Entrance",
-    "age": 45,
-    "severity": "ESI3",
-    "patientName": "Christopher",
-    "gender": "male",
-    "id": 62
-  },
-  //
-  // PAST HERE ARE THE SHIFT CHANGES AT 6 PM
-  //
-  // {
-  //   "name": "Nurse",
-  //   "type": "Triage Nurse",
-  //   "arrivalLocation": "Main Entrance",
-  //   "arrivalTick": 1728000,
-  //   //"arrivalTick": 116000,
-  //   "id": 63
-  // },
-  // {
-  //   "name": "Nurse",
-  //   "type": "Triage Nurse",
-  //   "arrivalLocation": "Main Entrance",
-  //   "arrivalTick": 1728000,
-  //   //"arrivalTick": 116000,
-  //   "id": 64
-  // },
-  // {
-  //   "name": "Tech",
-  //   "type": "Tech",
-  //   "arrivalLocation": "Main Entrance",
-  //   "arrivalTick": 1728000,
-  //   //"arrivalTick": 116000,
-  //   "id": 65
-  // },
-  // {
-  //   "name": "Tech",
-  //   "type": "Tech",
-  //   "arrivalLocation": "Main Entrance",
-  //   "arrivalTick": 1728000,
-  //   //"arrivalTick": 116000,
-  //   "id": 66
-  // },
-  // {
-  //   "name": "Nurse",
-  //   "type": "Nurse",
-  //   "arrivalLocation": "Main Entrance",
-  //   "arrivalTick": 1728000,
-  //   //"arrivalTick": 116000,
-  //   "id": 67
-  // },
-  // {
-  //   "name": "Nurse",
-  //   "type": "Nurse",
-  //   "arrivalLocation": "Main Entrance",
-  //   "arrivalTick": 1728000,
-  //   //"arrivalTick": 116000,
-  //   "id": 68
-  // },
-  // {
-  //   "name": "Nurse",
-  //   "type": "Greeter Nurse",
-  //   "arrivalLocation": "Main Entrance",
-  //   "arrivalTick": 1728000,
-  //   //"arrivalTick": 116000,
-  //   "id": 69
-  // },
-  // {
-  //   "name": "Tech",
-  //   "type": "CT",
-  //   "arrivalLocation": "Main Entrance",
-  //   "arrivalTick": 1728000,
-  //   //"arrivalTick": 116000,
-  //   "id": 70
-  // },
-  // {
-  //   "name": "Tech",
-  //   "type": "CT",
-  //   "arrivalLocation": "Main Entrance",
-  //   "arrivalTick": 1728000,
-  //   //"arrivalTick": 116000,
-  //   "id": 71
-  // },
-  // {
-  //   "name": "Tech",
-  //   "type": "Janitorial",
-  //   "arrivalLocation": "Main Entrance",
-  //   "arrivalTick": 1728000,
-  //   //"arrivalTick": 116000,
-  //   "id": 72
-  // },
-  // {
-  //   "name": "Tech",
-  //   "type": "Radiology",
-  //   "arrivalLocation": "Main Entrance",
-  //   "arrivalTick": 1728000,
-  //   //"arrivalTick": 116000,
-  //   "id": 73
-  // },
-  // // {
-  // //   "name": "Attending",
-  // //   "type": "Attending",
-  // //   "arrivalLocation": "Main Entrance",
-  // //   "arrivalTick": 1728000,
-  // //   "id": 42
-  // // },
-  // {
-  //   "name": "Resident",
-  //   "type": "Resident",
-  //   "arrivalLocation": "Main Entrance",
-  //   "arrivalTick": 1728000,
-  //   //"arrivalTick": 116000,
-  //   "id": 74
-  // },
-  // {
-  //   "name": "Resident",
-  //   "type": "Resident",
-  //   "arrivalLocation": "Main Entrance",
-  //   "arrivalTick": 1728000,
-  //   //"arrivalTick": 116000,
-  //   "id": 75
-  // },
-  // // {
-  // //   "name": "Pharmacist",
-  // //   "type": "Pharmacist",
-  // //   "arrivalLocation": "Main Entrance",
-  // //   "arrivalTick": 1728000,
-  // //   "id": 45
-  // // },
-  // {
-  //   "name": "Tech",
-  //   "type": "XRay",
-  //   "arrivalLocation": "Main Entrance",
-  //   "arrivalTick": 1728000,
-  //   //"arrivalTick": 116000,
-  //   "id": 76
-  // },
-  // {
-  //   "name": "Tech",
-  //   "type": "XRay",
-  //   "arrivalLocation": "Main Entrance",
-  //   "arrivalTick": 1728000,
-  //   //"arrivalTick": 116000,
-  //   "id": 77
-  // },
-  // {
-  //   "name": "Tech",
-  //   "type": "Phlebotomist",
-  //   "arrivalLocation": "Main Entrance",
-  //   "arrivalTick": 1728000,
-  //   //"arrivalTick": 116000,
-  //   "id": 78
-  // },
-  // {
-  //   "name": "Tech",
-  //   "type": "Phlebotomist",
-  //   "arrivalLocation": "Main Entrance",
-  //   "arrivalTick": 1728000,
-  //   //"arrivalTick": 116000,
-  //   "id": 79
-  // },
-  // {
-  //   "name": "Tech",
-  //   "type": "Phlebotomist",
-  //   "arrivalLocation": "Main Entrance",
-  //   "arrivalTick": 1728000,
-  //   //"arrivalTick": 116000,
-  //   "id": 80
-  // },
-  // {
-  //   "name": "Tech",
-  //   "type": "Phlebotomist",
-  //   "arrivalLocation": "Main Entrance",
-  //   "arrivalTick": 1728000,
-  //   //"arrivalTick": 116000,
-  //   "id": 81
-  // },
-  // PATIENTS ON SECOND HALF OF SIMULATION
-  {
-    "name": "patient",
-    "arrivalTick": 1728000,
-    "arrivalLocation": "Main Entrance",
-    "age": 28,
-    "severity": "ESI4",
-    "patientName": "Samantha",
-    "gender": "female",
-    "id": 63
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1742995,
-    "arrivalLocation": "Main Entrance",
-    "age": 5,
-    "severity": "ESI3",
-    "patientName": "Olivia",
-    "gender": "female",
-    "id": 64
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1767667,
-    "arrivalLocation": "Main Entrance",
-    "age": 60,
-    "severity": "ESI3",
-    "patientName": "Linda",
-    "gender": "female",
-    "id": 65
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1787065,
-    "arrivalLocation": "Main Entrance",
-    "age": 65,
-    "severity": "ESI3",
-    "patientName": "John",
-    "gender": "male",
-    "id": 66
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1852418,
-    "arrivalLocation": "Main Entrance",
-    "age": 24,
-    "severity": "ESI2",
-    "patientName": "Sarah",
-    "gender": "female",
-    "id": 67
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1872222,
-    "arrivalLocation": "Main Entrance",
-    "age": 37,
-    "severity": "ESI2",
-    "patientName": "Melissa",
-    "gender": "female",
-    "id": 68
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1912333,
-    "arrivalLocation": "Main Entrance",
-    "age": 22,
-    "severity": "ESI3",
-    "patientName": "Michael",
-    "gender": "male",
-    "id": 69
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1931397,
-    "arrivalLocation": "Main Entrance",
-    "age": 7,
-    "severity": "ESI3",
-    "patientName": "Isabella",
-    "gender": "female",
-    "id": 70
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1944564,
-    "arrivalLocation": "Main Entrance",
-    "age": 27,
-    "severity": "ESI3",
-    "patientName": "Jessica",
-    "gender": "female",
-    "id": 71
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 1970690,
-    "arrivalLocation": "Main Entrance",
-    "age": 71,
-    "severity": "ESI3",
-    "patientName": "James",
-    "gender": "male",
-    "id": 72
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 2012880,
-    "arrivalLocation": "Main Entrance",
-    "age": 66,
-    "severity": "ESI3",
-    "patientName": "Deborah",
-    "gender": "female",
-    "id": 73
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 2025201,
-    "arrivalLocation": "Main Entrance",
-    "age": 13,
-    "severity": "ESI3",
-    "patientName": "Michael",
-    "gender": "male",
-    "id": 74
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 2090241,
-    "arrivalLocation": "Main Entrance",
-    "age": 6,
-    "severity": "ESI2",
-    "patientName": "Emma",
-    "gender": "female",
-    "id": 75
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 2110772,
-    "arrivalLocation": "Main Entrance",
-    "age": 27,
-    "severity": "ESI3",
-    "patientName": "Andrew",
-    "gender": "male",
-    "id": 76
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 2138323,
-    "arrivalLocation": "Main Entrance",
-    "age": 12,
-    "severity": "ESI3",
-    "patientName": "Emily",
-    "gender": "female",
-    "id": 77
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 2159627,
-    "arrivalLocation": "Main Entrance",
-    "age": 20,
-    "severity": "ESI2",
-    "patientName": "Matthew",
-    "gender": "male",
-    "id": 78
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 2179797,
-    "arrivalLocation": "Main Entrance",
-    "age": 19,
-    "severity": "ESI3",
-    "patientName": "Sarah",
-    "gender": "female",
-    "id": 79
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 2186980,
-    "arrivalLocation": "Main Entrance",
-    "age": 30,
-    "severity": "ESI3",
-    "patientName": "Brittany",
-    "gender": "female",
-    "id": 80
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 2249359,
-    "arrivalLocation": "Main Entrance",
-    "age": 54,
-    "severity": "ESI3",
-    "patientName": "Mary",
-    "gender": "female",
-    "id": 81
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 2275397,
-    "arrivalLocation": "Main Entrance",
-    "age": 28,
-    "severity": "ESI2",
-    "patientName": "Samantha",
-    "gender": "female",
-    "id": 82
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 2301724,
-    "arrivalLocation": "Main Entrance",
-    "age": 80,
-    "severity": "ESI3",
-    "patientName": "John",
-    "gender": "male",
-    "id": 83
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 2322728,
-    "arrivalLocation": "Main Entrance",
-    "age": 51,
-    "severity": "ESI3",
-    "patientName": "Melissa",
-    "gender": "female",
-    "id": 84
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 2382157,
-    "arrivalLocation": "Main Entrance",
-    "age": 67,
-    "severity": "ESI3",
-    "patientName": "James",
-    "gender": "male",
-    "id": 85
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 2393101,
-    "arrivalLocation": "Main Entrance",
-    "age": 37,
-    "severity": "ESI3",
-    "patientName": "Christopher",
-    "gender": "male",
-    "id": 86
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 2438432,
-    "arrivalLocation": "Main Entrance",
-    "age": 51,
-    "severity": "ESI2",
-    "patientName": "Melissa",
-    "gender": "female",
-    "id": 87
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 2486298,
-    "arrivalLocation": "Main Entrance",
-    "age": 32,
-    "severity": "ESI3",
-    "patientName": "Amanda",
-    "gender": "female",
-    "id": 88
-  },
-  {
-    "name": "patient",
-    "arrivalTick": 2525020,
-    "arrivalLocation": "Main Entrance",
-    "age": 38,
-    "severity": "ESI2",
-    "patientName": "Melissa",
-    "gender": "female",
-    "id": 89
-  },
-]
