--- src/assets/arrivals/arrivalHospitalThesisCopy.js
@@ -1,2356 +1,0 @@
-export default [
-    {
-      "name": "patient",
-      "arrivalTick": 0,
-      "arrivalLocation": "Main Entrance",
-      "age": 34,
-      "severity": "ESI3",
-      "patientName": "Daniel",
-      "gender": "male",
-      "id": 0
-    },
-    // THESE ARE THE 6 AM NURSES
-    {
-      "name": "Nurse",
-      "type": "Triage Nurse",
-      "arrivalLocation": "Main Entrance",
-      "arrivalTick": 10,
-      "id": 1
-    },
-    {
-      "name": "Nurse",
-      "type": "Triage Nurse",
-      "arrivalLocation": "Main Entrance",
-      "arrivalTick": 10,
-      "id": 2
-    },
-    {
-      "name": "Tech",
-      "type": "Tech",
-      "arrivalLocation": "Main Entrance",
-      "arrivalTick": 10,
-      "id": 3
-    },
-    {
-      "name": "Tech",
-      "type": "Tech",
-      "arrivalLocation": "Main Entrance",
-      "arrivalTick": 10,
-      "id": 4
-    },
-    {
-      "name": "Nurse",
-      "type": "Nurse",
-      "arrivalLocation": "Main Entrance",
-      "arrivalTick": 10,
-      "id": 5
-    },
-    {
-      "name": "Nurse",
-      "type": "Nurse",
-      "arrivalLocation": "Main Entrance",
-      "arrivalTick": 10,
-      "id": 6
-    },
-    {
-      "name": "Nurse",
-      "type": "Greeter Nurse",
-      "arrivalLocation": "Main Entrance",
-      "arrivalTick": 10,
-      "id": 7
-    },
-    {
-      "name": "Tech",
-      "type": "CT",
-      "arrivalLocation": "Main Entrance",
-      "arrivalTick": 10,
-      "id": 8
-    },
-    {
-      "name": "Tech",
-      "type": "CT",
-      "arrivalLocation": "Main Entrance",
-      "arrivalTick": 10,
-      "id": 9
-    },
-    {
-      "name": "Tech",
-      "type": "Janitorial",
-      "arrivalLocation": "Main Entrance",
-      "arrivalTick": 10,
-      "id": 10
-    },
-    {
-      "name": "Tech",
-      "type": "Radiology",
-      "arrivalLocation": "Main Entrance",
-      "arrivalTick": 10,
-      "id": 11
-    },
-    {
-      "name": "Attending",
-      "type": "Attending",
-      "arrivalLocation": "Main Entrance",
-      "arrivalTick": 20,
-      "id": 12
-    },
-    {
-      "name": "Resident",
-      "type": "Resident",
-      "arrivalLocation": "Main Entrance",
-      "arrivalTick": 20,
-      "id": 13
-    },
-    {
-      "name": "Resident",
-      "type": "Resident",
-      "arrivalLocation": "Main Entrance",
-      "arrivalTick": 20,
-      "id": 14
-    },
-    {
-      "name": "Pharmacist",
-      "type": "Pharmacist",
-      "arrivalLocation": "Main Entrance",
-      "arrivalTick": 30,
-      "id": 15
-    },
-    {
-      "name": "Tech",
-      "type": "XRay",
-      "arrivalLocation": "Main Entrance",
-      "arrivalTick": 40,
-      "id": 16
-    },
-    {
-      "name": "Tech",
-      "type": "XRay",
-      "arrivalLocation": "Main Entrance",
-      "arrivalTick": 40,
-      "id": 17
-    },
-    {
-      "name": "Tech",
-      "type": "Phlebotomist",
-      "arrivalLocation": "Main Entrance",
-      "arrivalTick": 50,
-      "id": 18
-    },
-    {
-      "name": "Tech",
-      "type": "Phlebotomist",
-      "arrivalLocation": "Main Entrance",
-      "arrivalTick": 50,
-      "id": 19
-    },
-    {
-      "name": "Tech",
-      "type": "Phlebotomist",
-      "arrivalLocation": "Main Entrance",
-      "arrivalTick": 50,
-      "id": 20
-    },
-    {
-      "name": "Tech",
-      "type": "Phlebotomist",
-      "arrivalLocation": "Main Entrance",
-      "arrivalTick": 50,
-      "id": 21
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 30057,
-      "arrivalLocation": "Main Entrance",
-      "age": 39,
-      "severity": "ESI3",
-      "patientName": "James",
-      "gender": "male",
-      "id": 22
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 46016,
-      "arrivalLocation": "Main Entrance",
-      "age": 12,
-      "severity": "ESI3",
-      "patientName": "Emma",
-      "gender": "female",
-      "id": 23
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 77178,
-      "arrivalLocation": "Main Entrance",
-      "age": 69,
-      // "severity": "ESI3",
-      "severity": "ESI3",
-      "patientName": "Susan",
-      "gender": "female",
-      "id": 24
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 105239,
-      "arrivalLocation": "Main Entrance",
-      "age": 12,
-      "severity": "ESI3",
-      "patientName": "Emma",
-      "gender": "female",
-      "id": 25
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 105755,
-      "arrivalLocation": "Main Entrance",
-      "age": 62,
-      "severity": "ESI2",
-      "patientName": "David",
-      "gender": "male",
-      "id": 26
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 108000,
-      "arrivalLocation": "Main Entrance",
-      "age": 17,
-      "severity": "ESI3",
-      "patientName": "Emily",
-      "gender": "female",
-      "id": 27
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 128609,
-      "arrivalLocation": "Main Entrance",
-      "age": 38,
-      "severity": "ESI3",
-      "patientName": "Michael",
-      "gender": "male",
-      "id": 28
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 131256,
-      "arrivalLocation": "Main Entrance",
-      "age": 30,
-      "severity": "ESI3",
-      "patientName": "David",
-      "gender": "male",
-      "id": 29
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 132392,
-      "arrivalLocation": "Main Entrance",
-      "age": 79,
-      "severity": "ESI3",
-      "patientName": "James",
-      "gender": "male",
-      "id": 30
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 162719,
-      "arrivalLocation": "Main Entrance",
-      "age": 27,
-      "severity": "ESI3",
-      "patientName": "Michael",
-      "gender": "male",
-      "id": 31
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 216000,
-      "arrivalLocation": "Main Entrance",
-      "age": 57,
-      "severity": "ESI3",
-      "patientName": "Robert",
-      "gender": "male",
-      "id": 32
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 224438,
-      "arrivalLocation": "Main Entrance",
-      "age": 57,
-      "severity": "ESI3",
-      "patientName": "Karen",
-      "gender": "female",
-      "id": 33
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 234701,
-      "arrivalLocation": "Main Entrance",
-      "age": 11,
-      "severity": "ESI3",
-      "patientName": "Ethan",
-      "gender": "male",
-      "id": 34
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 301995,
-      "arrivalLocation": "Main Entrance",
-      "age": 49,
-      "severity": "ESI2",
-      "patientName": "Amy",
-      "gender": "female",
-      "id": 35
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 324000,
-      "arrivalLocation": "Main Entrance",
-      "age": 21,
-      "severity": "ESI3",
-      "patientName": "Michael",
-      "gender": "male",
-      "id": 36
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 370377,
-      "arrivalLocation": "Main Entrance",
-      "age": 60,
-      "severity": "ESI3",
-      "patientName": "Linda",
-      "gender": "female",
-      "id": 37
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 395765,
-      "arrivalLocation": "Main Entrance",
-      "age": 25,
-      "severity": "ESI2",
-      "patientName": "Joshua",
-      "gender": "male",
-      "id": 38
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 402231,
-      "arrivalLocation": "Main Entrance",
-      "age": 14,
-      "severity": "ESI3",
-      "patientName": "Matthew",
-      "gender": "male",
-      "id": 39
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 432000,
-      "arrivalLocation": "Ambulance Entrance",
-      "age": 39,
-      "severity": "ESI1",
-      "patientName": "David",
-      "gender": "male",
-      "id": 40
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 439500,
-      "arrivalLocation": "Main Entrance",
-      "age": 52,
-      "severity": "ESI3",
-      "patientName": "Robert",
-      "gender": "male",
-      "id": 41
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 453747,
-      "arrivalLocation": "Main Entrance",
-      "age": 4,
-      "severity": "ESI2",
-      "patientName": "Sophia",
-      "gender": "female",
-      "id": 42
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 473877,
-      "arrivalLocation": "Main Entrance",
-      "age": 1,
-      "severity": "ESI2",
-      "patientName": "Sophia",
-      "gender": "female",
-      "id": 43
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 540000,
-      "arrivalLocation": "Main Entrance",
-      "age": 9,
-      "severity": "ESI2",
-      "patientName": "Olivia",
-      "gender": "female",
-      "id": 44
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 565811,
-      "arrivalLocation": "Main Entrance",
-      "age": 62,
-      "severity": "ESI2",
-      "patientName": "David",
-      "gender": "male",
-      "id": 45
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 578782,
-      "arrivalLocation": "Main Entrance",
-      "age": 12,
-      "severity": "ESI2",
-      "patientName": "Joshua",
-      "gender": "male",
-      "id": 46
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 612374,
-      "arrivalLocation": "Main Entrance",
-      "age": 27,
-      "severity": "ESI3",
-      "patientName": "Sarah",
-      "gender": "female",
-      "id": 47
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 648000,
-      "arrivalLocation": "Main Entrance",
-      "age": 2,
-      "severity": "ESI2",
-      "patientName": "Ava",
-      "gender": "female",
-      "id": 48
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 650188,
-      "arrivalLocation": "Main Entrance",
-      "age": 46,
-      "severity": "ESI3",
-      "patientName": "James",
-      "gender": "male",
-      "id": 49
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 674341,
-      "arrivalLocation": "Main Entrance",
-      "age": 1,
-      "severity": "ESI2",
-      "patientName": "Ava",
-      "gender": "female",
-      "id": 50
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 709163,
-      "arrivalLocation": "Main Entrance",
-      "age": 37,
-      "severity": "ESI3",
-      "patientName": "Sarah",
-      "gender": "female",
-      "id": 51
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 756000,
-      "arrivalLocation": "Main Entrance",
-      "age": 52,
-      "severity": "ESI2",
-      "patientName": "Kimberly",
-      "gender": "female",
-      "id": 52
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 770109,
-      "arrivalLocation": "Main Entrance",
-      "age": 25,
-      "severity": "ESI2",
-      "patientName": "Michael",
-      "gender": "male",
-      "id": 53
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 809758,
-      "arrivalLocation": "Main Entrance",
-      "age": 48,
-      "severity": "ESI3",
-      "patientName": "Jennifer",
-      "gender": "female",
-      "id": 54
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 813603,
-      "arrivalLocation": "Main Entrance",
-      "age": 53,
-      "severity": "ESI3",
-      "patientName": "Michael",
-      "gender": "male",
-      "id": 55
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 846588,
-      "arrivalLocation": "Main Entrance",
-      "age": 65,
-      "severity": "ESI3",
-      "patientName": "Michael",
-      "gender": "male",
-      "id": 56
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 864000,
-      "arrivalLocation": "Main Entrance",
-      "age": 18,
-      "severity": "ESI3",
-      "patientName": "Michael",
-      "gender": "male",
-      "id": 57
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 876574,
-      "arrivalLocation": "Main Entrance",
-      "age": 32,
-      "severity": "ESI3",
-      "patientName": "Jessica",
-      "gender": "female",
-      "id": 58
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 885499,
-      "arrivalLocation": "Main Entrance",
-      "age": 11,
-      "severity": "ESI2",
-      "patientName": "Emily",
-      "gender": "female",
-      "id": 59
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 888214,
-      "arrivalLocation": "Main Entrance",
-      "age": 23,
-      "severity": "ESI3",
-      "patientName": "Joshua",
-      "gender": "male",
-      "id": 60
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 930902,
-      "arrivalLocation": "Main Entrance",
-      "age": 35,
-      "severity": "ESI2",
-      "patientName": "Jennifer",
-      "gender": "female",
-      "id": 61
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 938194,
-      "arrivalLocation": "Main Entrance",
-      "age": 10,
-      "severity": "ESI3",
-      "patientName": "Olivia",
-      "gender": "female",
-      "id": 62
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 970236,
-      "arrivalLocation": "Main Entrance",
-      "age": 10,
-      "severity": "ESI3",
-      "patientName": "Isabella",
-      "gender": "female",
-      "id": 63
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 972000,
-      "arrivalLocation": "Main Entrance",
-      "age": 18,
-      "severity": "ESI2",
-      "patientName": "Jacob",
-      "gender": "male",
-      "id": 64
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 974719,
-      "arrivalLocation": "Main Entrance",
-      "age": 18,
-      "severity": "ESI2",
-      "patientName": "Michael",
-      "gender": "male",
-      "id": 65
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1001661,
-      "arrivalLocation": "Main Entrance",
-      "age": 65,
-      "severity": "ESI2",
-      "patientName": "Linda",
-      "gender": "female",
-      "id": 66
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1018100,
-      "arrivalLocation": "Main Entrance",
-      "age": 14,
-      "severity": "ESI2",
-      "patientName": "Matthew",
-      "gender": "male",
-      "id": 67
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1024863,
-      "arrivalLocation": "Main Entrance",
-      "age": 70,
-      "severity": "ESI3",
-      "patientName": "John",
-      "gender": "male",
-      "id": 68
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1030423,
-      "arrivalLocation": "Main Entrance",
-      "age": 36,
-      "severity": "ESI3",
-      "patientName": "Michael",
-      "gender": "male",
-      "id": 69
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1033598,
-      "arrivalLocation": "Main Entrance",
-      "age": 59,
-      "severity": "ESI3",
-      "patientName": "Donna",
-      "gender": "female",
-      "id": 70
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1037859,
-      "arrivalLocation": "Main Entrance",
-      "age": 55,
-      "severity": "ESI3",
-      "patientName": "John",
-      "gender": "male",
-      "id": 71
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1045057,
-      "arrivalLocation": "Main Entrance",
-      "age": 43,
-      "severity": "ESI2",
-      "patientName": "Michael",
-      "gender": "male",
-      "id": 72
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1062250,
-      "arrivalLocation": "Main Entrance",
-      "age": 42,
-      "severity": "ESI3",
-      "patientName": "Michael",
-      "gender": "male",
-      "id": 73
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1080000,
-      "arrivalLocation": "Main Entrance",
-      "age": 45,
-      "severity": "ESI3",
-      "patientName": "Christopher",
-      "gender": "male",
-      "id": 74
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1084534,
-      "arrivalLocation": "Main Entrance",
-      "age": 94,
-      "severity": "ESI2",
-      "patientName": "Mary",
-      "gender": "female",
-      "id": 75
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1094277,
-      "arrivalLocation": "Main Entrance",
-      "age": 20,
-      "severity": "ESI3",
-      "patientName": "Sarah",
-      "gender": "female",
-      "id": 76
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1115762,
-      "arrivalLocation": "Main Entrance",
-      "age": 13,
-      "severity": "ESI3",
-      "patientName": "Emily",
-      "gender": "female",
-      "id": 77
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1120783,
-      "arrivalLocation": "Main Entrance",
-      "age": 44,
-      "severity": "ESI2",
-      "patientName": "David",
-      "gender": "male",
-      "id": 78
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1121791,
-      "arrivalLocation": "Main Entrance",
-      "age": 38,
-      "severity": "ESI2",
-      "patientName": "Jennifer",
-      "gender": "female",
-      "id": 79
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1143590,
-      "arrivalLocation": "Main Entrance",
-      "age": 43,
-      "severity": "ESI3",
-      "patientName": "David",
-      "gender": "male",
-      "id": 80
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1143865,
-      "arrivalLocation": "Main Entrance",
-      "age": 30,
-      "severity": "ESI3",
-      "patientName": "Joshua",
-      "gender": "male",
-      "id": 81
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1169077,
-      "arrivalLocation": "Main Entrance",
-      "age": 1,
-      "severity": "ESI3",
-      "patientName": "Oliver",
-      "gender": "male",
-      "id": 82
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1183285,
-      "arrivalLocation": "Main Entrance",
-      "age": 45,
-      "severity": "ESI3",
-      "patientName": "Angela",
-      "gender": "female",
-      "id": 83
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1184194,
-      "arrivalLocation": "Main Entrance",
-      "age": 22,
-      "severity": "ESI3",
-      "patientName": "Christopher",
-      "gender": "male",
-      "id": 84
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1188000,
-      "arrivalLocation": "Main Entrance",
-      "age": 44,
-      "severity": "ESI3",
-      "patientName": "David",
-      "gender": "male",
-      "id": 85
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1188757,
-      "arrivalLocation": "Main Entrance",
-      "age": 22,
-      "severity": "ESI2",
-      "patientName": "Hannah",
-      "gender": "female",
-      "id": 86
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1198188,
-      "arrivalLocation": "Main Entrance",
-      "age": 56,
-      "severity": "ESI3",
-      "patientName": "Lisa",
-      "gender": "female",
-      "id": 87
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1206684,
-      "arrivalLocation": "Main Entrance",
-      "age": 9,
-      "severity": "ESI2",
-      "patientName": "Olivia",
-      "gender": "female",
-      "id": 88
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1212254,
-      "arrivalLocation": "Main Entrance",
-      "age": 36,
-      "severity": "ESI2",
-      "patientName": "Jessica",
-      "gender": "female",
-      "id": 89
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1236019,
-      "arrivalLocation": "Ambulance Entrance",
-      "age": 2,
-      "severity": "ESI1",
-      "patientName": "Isabella",
-      "gender": "female",
-      "id": 90
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1237691,
-      "arrivalLocation": "Main Entrance",
-      "age": 12,
-      "severity": "ESI2",
-      "patientName": "Isabella",
-      "gender": "female",
-      "id": 91
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1250087,
-      "arrivalLocation": "Main Entrance",
-      "age": 3,
-      "severity": "ESI2",
-      "patientName": "Ava",
-      "gender": "female",
-      "id": 92
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1252546,
-      "arrivalLocation": "Main Entrance",
-      "age": 47,
-      "severity": "ESI2",
-      "patientName": "Kimberly",
-      "gender": "female",
-      "id": 93
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1264071,
-      "arrivalLocation": "Main Entrance",
-      "age": 23,
-      "severity": "ESI3",
-      "patientName": "Michael",
-      "gender": "male",
-      "id": 94
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1264207,
-      "arrivalLocation": "Main Entrance",
-      "age": 15,
-      "severity": "ESI3",
-      "patientName": "Michael",
-      "gender": "male",
-      "id": 95
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1276884,
-      "arrivalLocation": "Main Entrance",
-      "age": 38,
-      "severity": "ESI3",
-      "patientName": "Matthew",
-      "gender": "male",
-      "id": 96
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1296000,
-      "arrivalLocation": "Main Entrance",
-      "age": 55,
-      "severity": "ESI2",
-      "patientName": "Mary",
-      "gender": "female",
-      "id": 97
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1299128,
-      "arrivalLocation": "Main Entrance",
-      "age": 8,
-      "severity": "ESI3",
-      "patientName": "Emma",
-      "gender": "female",
-      "id": 98
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1299367,
-      "arrivalLocation": "Main Entrance",
-      "age": 14,
-      "severity": "ESI2",
-      "patientName": "Matthew",
-      "gender": "male",
-      "id": 99
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1311043,
-      "arrivalLocation": "Main Entrance",
-      "age": 66,
-      "severity": "ESI3",
-      "patientName": "Robert",
-      "gender": "male",
-      "id": 100
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1325234,
-      "arrivalLocation": "Main Entrance",
-      "age": 54,
-      "severity": "ESI3",
-      "patientName": "Mary",
-      "gender": "female",
-      "id": 101
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1327346,
-      "arrivalLocation": "Main Entrance",
-      "age": 84,
-      "severity": "ESI2",
-      "patientName": "William",
-      "gender": "male",
-      "id": 102
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1332409,
-      "arrivalLocation": "Main Entrance",
-      "age": 2,
-      "severity": "ESI3",
-      "patientName": "James",
-      "gender": "male",
-      "id": 103
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1340022,
-      "arrivalLocation": "Main Entrance",
-      "age": 44,
-      "severity": "ESI3",
-      "patientName": "Jennifer",
-      "gender": "female",
-      "id": 104
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1365756,
-      "arrivalLocation": "Main Entrance",
-      "age": 71,
-      "severity": "ESI3",
-      "patientName": "Patricia",
-      "gender": "female",
-      "id": 105
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1368891,
-      "arrivalLocation": "Main Entrance",
-      "age": 9,
-      "severity": "ESI3",
-      "patientName": "Jayden",
-      "gender": "male",
-      "id": 106
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1382676,
-      "arrivalLocation": "Main Entrance",
-      "age": 1,
-      "severity": "ESI2",
-      "patientName": "Emma",
-      "gender": "female",
-      "id": 107
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1394895,
-      "arrivalLocation": "Main Entrance",
-      "age": 40,
-      "severity": "ESI3",
-      "patientName": "Jennifer",
-      "gender": "female",
-      "id": 108
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1404000,
-      "arrivalLocation": "Main Entrance",
-      "age": 34,
-      "severity": "ESI3",
-      "patientName": "Daniel",
-      "gender": "male",
-      "id": 109
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1404296,
-      "arrivalLocation": "Main Entrance",
-      "age": 56,
-      "severity": "ESI3",
-      "patientName": "Karen",
-      "gender": "female",
-      "id": 110
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1412097,
-      "arrivalLocation": "Main Entrance",
-      "age": 59,
-      "severity": "ESI2",
-      "patientName": "Donna",
-      "gender": "female",
-      "id": 111
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1416461,
-      "arrivalLocation": "Main Entrance",
-      "age": 66,
-      "severity": "ESI3",
-      "patientName": "Robert",
-      "gender": "male",
-      "id": 112
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1450526,
-      "arrivalLocation": "Main Entrance",
-      "age": 40,
-      "severity": "ESI3",
-      "patientName": "Melissa",
-      "gender": "female",
-      "id": 113
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1457151,
-      "arrivalLocation": "Main Entrance",
-      "age": 67,
-      "severity": "ESI2",
-      "patientName": "John",
-      "gender": "male",
-      "id": 114
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1457562,
-      "arrivalLocation": "Main Entrance",
-      "age": 29,
-      "severity": "ESI3",
-      "patientName": "Christopher",
-      "gender": "male",
-      "id": 115
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1460411,
-      "arrivalLocation": "Main Entrance",
-      "age": 69,
-      "severity": "ESI3",
-      "patientName": "Susan",
-      "gender": "female",
-      "id": 116
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1469658,
-      "arrivalLocation": "Main Entrance",
-      "age": 41,
-      "severity": "ESI3",
-      "patientName": "Jennifer",
-      "gender": "female",
-      "id": 117
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1475659,
-      "arrivalLocation": "Main Entrance",
-      "age": 75,
-      "severity": "ESI3",
-      "patientName": "James",
-      "gender": "male",
-      "id": 118
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1497980,
-      "arrivalLocation": "Main Entrance",
-      "age": 4,
-      "severity": "ESI2",
-      "patientName": "Mason",
-      "gender": "male",
-      "id": 119
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1506525,
-      "arrivalLocation": "Main Entrance",
-      "age": 16,
-      "severity": "ESI2",
-      "patientName": "Hannah",
-      "gender": "female",
-      "id": 120
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1512000,
-      "arrivalLocation": "Ambulance Entrance",
-      "age": 14,
-      "severity": "ESI1",
-      "patientName": "Emma",
-      "gender": "female",
-      "id": 121
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1531543,
-      "arrivalLocation": "Main Entrance",
-      "age": 19,
-      "severity": "ESI3",
-      "patientName": "Madison",
-      "gender": "female",
-      "id": 122
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1535291,
-      "arrivalLocation": "Main Entrance",
-      "age": 35,
-      "severity": "ESI3",
-      "patientName": "Jessica",
-      "gender": "female",
-      "id": 123
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1538080,
-      "arrivalLocation": "Main Entrance",
-      "age": 72,
-      "severity": "ESI3",
-      "patientName": "Barbara",
-      "gender": "female",
-      "id": 124
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1538977,
-      "arrivalLocation": "Main Entrance",
-      "age": 26,
-      "severity": "ESI3",
-      "patientName": "Tyler",
-      "gender": "male",
-      "id": 125
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1539842,
-      "arrivalLocation": "Main Entrance",
-      "age": 6,
-      "severity": "ESI2",
-      "patientName": "Isabella",
-      "gender": "female",
-      "id": 126
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1542210,
-      "arrivalLocation": "Main Entrance",
-      "age": 66,
-      "severity": "ESI2",
-      "patientName": "Michael",
-      "gender": "male",
-      "id": 127
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1568116,
-      "arrivalLocation": "Main Entrance",
-      "age": 52,
-      "severity": "ESI3",
-      "patientName": "Mary",
-      "gender": "female",
-      "id": 128
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1570853,
-      "arrivalLocation": "Main Entrance",
-      "age": 43,
-      "severity": "ESI3",
-      "patientName": "James",
-      "gender": "male",
-      "id": 129
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1577255,
-      "arrivalLocation": "Main Entrance",
-      "age": 50,
-      "severity": "ESI2",
-      "patientName": "Melissa",
-      "gender": "female",
-      "id": 130
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1581303,
-      "arrivalLocation": "Main Entrance",
-      "age": 5,
-      "severity": "ESI2",
-      "patientName": "Mason",
-      "gender": "male",
-      "id": 131
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1620000,
-      "arrivalLocation": "Main Entrance",
-      "age": 23,
-      "severity": "ESI3",
-      "patientName": "Matthew",
-      "gender": "male",
-      "id": 132
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1638395,
-      "arrivalLocation": "Main Entrance",
-      "age": 74,
-      "severity": "ESI3",
-      "patientName": "Carol",
-      "gender": "female",
-      "id": 133
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1645655,
-      "arrivalLocation": "Main Entrance",
-      "age": 34,
-      "severity": "ESI3",
-      "patientName": "Jennifer",
-      "gender": "female",
-      "id": 134
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1653377,
-      "arrivalLocation": "Main Entrance",
-      "age": 3,
-      "severity": "ESI2",
-      "patientName": "Isabella",
-      "gender": "female",
-      "id": 135
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1671995,
-      "arrivalLocation": "Main Entrance",
-      "age": 14,
-      "severity": "ESI3",
-      "patientName": "Madison",
-      "gender": "female",
-      "id": 136
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1680865,
-      "arrivalLocation": "Main Entrance",
-      "age": 55,
-      "severity": "ESI3",
-      "patientName": "David",
-      "gender": "male",
-      "id": 137
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1690977,
-      "arrivalLocation": "Main Entrance",
-      "age": 12,
-      "severity": "ESI3",
-      "patientName": "Ethan",
-      "gender": "male",
-      "id": 138
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1701194,
-      "arrivalLocation": "Main Entrance",
-      "age": 39,
-      "severity": "ESI3",
-      "patientName": "Christopher",
-      "gender": "male",
-      "id": 139
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1702307,
-      "arrivalLocation": "Main Entrance",
-      "age": 45,
-      "severity": "ESI2",
-      "patientName": "Amy",
-      "gender": "female",
-      "id": 140
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1705437,
-      "arrivalLocation": "Main Entrance",
-      "age": 17,
-      "severity": "ESI2",
-      "patientName": "Michael",
-      "gender": "male",
-      "id": 141
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1713298,
-      "arrivalLocation": "Main Entrance",
-      "age": 45,
-      "severity": "ESI3",
-      "patientName": "Christopher",
-      "gender": "male",
-      "id": 142
-    },
-    //
-    // PAST HERE ARE THE SHIFT CHANGES AT 6 PM
-    //
-    // {
-    //   "name": "Nurse",
-    //   "type": "Triage Nurse",
-    //   "arrivalLocation": "Main Entrance",
-    //   //"arrivalTick": 1728000,
-    //   "arrivalTick": 116000,
-    //   "id": 143
-    // },
-    // {
-    //   "name": "Nurse",
-    //   "type": "Triage Nurse",
-    //   "arrivalLocation": "Main Entrance",
-    //   //"arrivalTick": 1728000,
-    //   "arrivalTick": 116000,
-    //   "id": 144
-    // },
-    // {
-    //   "name": "Tech",
-    //   "type": "Tech",
-    //   "arrivalLocation": "Main Entrance",
-    //   //"arrivalTick": 1728000,
-    //   "arrivalTick": 116000,
-    //   "id": 145
-    // },
-    // {
-    //   "name": "Tech",
-    //   "type": "Tech",
-    //   "arrivalLocation": "Main Entrance",
-    //   //"arrivalTick": 1728000,
-    //   "arrivalTick": 116000,
-    //   "id": 146
-    // },
-    // {
-    //   "name": "Nurse",
-    //   "type": "Nurse",
-    //   "arrivalLocation": "Main Entrance",
-    //   //"arrivalTick": 1728000,
-    //   "arrivalTick": 116000,
-    //   "id": 147
-    // },
-    // {
-    //   "name": "Nurse",
-    //   "type": "Nurse",
-    //   "arrivalLocation": "Main Entrance",
-    //   //"arrivalTick": 1728000,
-    //   "arrivalTick": 116000,
-    //   "id": 148
-    // },
-    // {
-    //   "name": "Nurse",
-    //   "type": "Greeter Nurse",
-    //   "arrivalLocation": "Main Entrance",
-    //   //"arrivalTick": 1728000,
-    //   "arrivalTick": 116000,
-    //   "id": 149
-    // },
-    // {
-    //   "name": "Tech",
-    //   "type": "CT",
-    //   "arrivalLocation": "Main Entrance",
-    //   //"arrivalTick": 1728000,
-    //   "arrivalTick": 116000,
-    //   "id": 150
-    // },
-    // {
-    //   "name": "Tech",
-    //   "type": "CT",
-    //   "arrivalLocation": "Main Entrance",
-    //   //"arrivalTick": 1728000,
-    //   "arrivalTick": 116000,
-    //   "id": 151
-    // },
-    // {
-    //   "name": "Tech",
-    //   "type": "Janitorial",
-    //   "arrivalLocation": "Main Entrance",
-    //   //"arrivalTick": 1728000,
-    //   "arrivalTick": 116000,
-    //   "id": 152
-    // },
-    // {
-    //   "name": "Tech",
-    //   "type": "Radiology",
-    //   "arrivalLocation": "Main Entrance",
-    //   //"arrivalTick": 1728000,
-    //   "arrivalTick": 116000,
-    //   "id": 153
-    // },
-    // // {
-    // //   "name": "Attending",
-    // //   "type": "Attending",
-    // //   "arrivalLocation": "Main Entrance",
-    // //   "arrivalTick": 1728000,
-    // //   "id": 42
-    // // },
-    // {
-    //   "name": "Resident",
-    //   "type": "Resident",
-    //   "arrivalLocation": "Main Entrance",
-    //   //"arrivalTick": 1728000,
-    //   "arrivalTick": 116000,
-    //   "id": 154
-    // },
-    // {
-    //   "name": "Resident",
-    //   "type": "Resident",
-    //   "arrivalLocation": "Main Entrance",
-    //   //"arrivalTick": 1728000,
-    //   "arrivalTick": 116000,
-    //   "id": 155
-    // },
-    // // {
-    // //   "name": "Pharmacist",
-    // //   "type": "Pharmacist",
-    // //   "arrivalLocation": "Main Entrance",
-    // //   "arrivalTick": 1728000,
-    // //   "id": 45
-    // // },
-    // {
-    //   "name": "Tech",
-    //   "type": "XRay",
-    //   "arrivalLocation": "Main Entrance",
-    //   //"arrivalTick": 1728000,
-    //   "arrivalTick": 116000,
-    //   "id": 156
-    // },
-    // {
-    //   "name": "Tech",
-    //   "type": "XRay",
-    //   "arrivalLocation": "Main Entrance",
-    //   //"arrivalTick": 1728000,
-    //   "arrivalTick": 116000,
-    //   "id": 157
-    // },
-    // {
-    //   "name": "Tech",
-    //   "type": "Phlebotomist",
-    //   "arrivalLocation": "Main Entrance",
-    //   //"arrivalTick": 1728000,
-    //   "arrivalTick": 116000,
-    //   "id": 158
-    // },
-    // {
-    //   "name": "Tech",
-    //   "type": "Phlebotomist",
-    //   "arrivalLocation": "Main Entrance",
-    //   //"arrivalTick": 1728000,
-    //   "arrivalTick": 116000,
-    //   "id": 159
-    // },
-    // {
-    //   "name": "Tech",
-    //   "type": "Phlebotomist",
-    //   "arrivalLocation": "Main Entrance",
-    //   //"arrivalTick": 1728000,
-    //   "arrivalTick": 116000,
-    //   "id": 160
-    // },
-    // {
-    //   "name": "Tech",
-    //   "type": "Phlebotomist",
-    //   "arrivalLocation": "Main Entrance",
-    //   //"arrivalTick": 1728000,
-    //   "arrivalTick": 116000,
-    //   "id": 161
-    // },
-    // PATIENTS ON SECOND HALF OF SIMULATION
-    {
-      "name": "patient",
-      "arrivalTick": 1728000,
-      "arrivalLocation": "Main Entrance",
-      "age": 28,
-      "severity": "ESI4",
-      "patientName": "Samantha",
-      "gender": "female",
-      "id": 162
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1732198,
-      "arrivalLocation": "Main Entrance",
-      "age": 7,
-      "severity": "ESI3",
-      "patientName": "Jacob",
-      "gender": "male",
-      "id": 163
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1742995,
-      "arrivalLocation": "Main Entrance",
-      "age": 5,
-      "severity": "ESI3",
-      "patientName": "Olivia",
-      "gender": "female",
-      "id": 164
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1756939,
-      "arrivalLocation": "Main Entrance",
-      "age": 80,
-      "severity": "ESI3",
-      "patientName": "John",
-      "gender": "male",
-      "id": 165
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1764036,
-      "arrivalLocation": "Ambulance Entrance",
-      "age": 36,
-      "severity": "ESI1",
-      "patientName": "Christopher",
-      "gender": "male",
-      "id": 166
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1767233,
-      "arrivalLocation": "Main Entrance",
-      "age": 55,
-      "severity": "ESI2",
-      "patientName": "Mary",
-      "gender": "female",
-      "id": 167
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1767667,
-      "arrivalLocation": "Main Entrance",
-      "age": 60,
-      "severity": "ESI3",
-      "patientName": "Linda",
-      "gender": "female",
-      "id": 168
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1769775,
-      "arrivalLocation": "Main Entrance",
-      "age": 38,
-      "severity": "ESI3",
-      "patientName": "Jessica",
-      "gender": "female",
-      "id": 169
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1787065,
-      "arrivalLocation": "Main Entrance",
-      "age": 65,
-      "severity": "ESI3",
-      "patientName": "John",
-      "gender": "male",
-      "id": 170
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1812460,
-      "arrivalLocation": "Main Entrance",
-      "age": 45,
-      "severity": "ESI3",
-      "patientName": "James",
-      "gender": "male",
-      "id": 171
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1819639,
-      "arrivalLocation": "Main Entrance",
-      "age": 47,
-      "severity": "ESI3",
-      "patientName": "David",
-      "gender": "male",
-      "id": 172
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1836000,
-      "arrivalLocation": "Main Entrance",
-      "age": 14,
-      "severity": "ESI3",
-      "patientName": "Olivia",
-      "gender": "female",
-      "id": 173
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1852418,
-      "arrivalLocation": "Main Entrance",
-      "age": 24,
-      "severity": "ESI2",
-      "patientName": "Sarah",
-      "gender": "female",
-      "id": 174
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1859600,
-      "arrivalLocation": "Main Entrance",
-      "age": 77,
-      "severity": "ESI3",
-      "patientName": "James",
-      "gender": "male",
-      "id": 175
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1872222,
-      "arrivalLocation": "Main Entrance",
-      "age": 37,
-      "severity": "ESI2",
-      "patientName": "Melissa",
-      "gender": "female",
-      "id": 176
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1873730,
-      "arrivalLocation": "Main Entrance",
-      "age": 24,
-      "severity": "ESI3",
-      "patientName": "Sarah",
-      "gender": "female",
-      "id": 177
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1877474,
-      "arrivalLocation": "Main Entrance",
-      "age": 63,
-      "severity": "ESI3",
-      "patientName": "David",
-      "gender": "male",
-      "id": 178
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1908749,
-      "arrivalLocation": "Main Entrance",
-      "age": 83,
-      "severity": "ESI2",
-      "patientName": "John",
-      "gender": "male",
-      "id": 179
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1912333,
-      "arrivalLocation": "Main Entrance",
-      "age": 22,
-      "severity": "ESI3",
-      "patientName": "Michael",
-      "gender": "male",
-      "id": 180
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1915462,
-      "arrivalLocation": "Main Entrance",
-      "age": 78,
-      "severity": "ESI3",
-      "patientName": "Richard",
-      "gender": "male",
-      "id": 181
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1931397,
-      "arrivalLocation": "Main Entrance",
-      "age": 7,
-      "severity": "ESI3",
-      "patientName": "Isabella",
-      "gender": "female",
-      "id": 182
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1934041,
-      "arrivalLocation": "Main Entrance",
-      "age": 63,
-      "severity": "ESI2",
-      "patientName": "David",
-      "gender": "male",
-      "id": 183
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1940033,
-      "arrivalLocation": "Main Entrance",
-      "age": 58,
-      "severity": "ESI3",
-      "patientName": "Linda",
-      "gender": "female",
-      "id": 184
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1944000,
-      "arrivalLocation": "Main Entrance",
-      "age": 19,
-      "severity": "ESI3",
-      "patientName": "Joshua",
-      "gender": "male",
-      "id": 185
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1944564,
-      "arrivalLocation": "Main Entrance",
-      "age": 27,
-      "severity": "ESI3",
-      "patientName": "Jessica",
-      "gender": "female",
-      "id": 186
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1968443,
-      "arrivalLocation": "Main Entrance",
-      "age": 55,
-      "severity": "ESI3",
-      "patientName": "John",
-      "gender": "male",
-      "id": 187
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1970690,
-      "arrivalLocation": "Main Entrance",
-      "age": 71,
-      "severity": "ESI3",
-      "patientName": "James",
-      "gender": "male",
-      "id": 188
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1971092,
-      "arrivalLocation": "Main Entrance",
-      "age": 46,
-      "severity": "ESI3",
-      "patientName": "Lisa",
-      "gender": "female",
-      "id": 189
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1991298,
-      "arrivalLocation": "Main Entrance",
-      "age": 46,
-      "severity": "ESI3",
-      "patientName": "Michelle",
-      "gender": "female",
-      "id": 190
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 1997652,
-      "arrivalLocation": "Main Entrance",
-      "age": 48,
-      "severity": "ESI3",
-      "patientName": "Jennifer",
-      "gender": "female",
-      "id": 191
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2012880,
-      "arrivalLocation": "Main Entrance",
-      "age": 66,
-      "severity": "ESI3",
-      "patientName": "Deborah",
-      "gender": "female",
-      "id": 192
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2020876,
-      "arrivalLocation": "Main Entrance",
-      "age": 34,
-      "severity": "ESI2",
-      "patientName": "Christopher",
-      "gender": "male",
-      "id": 193
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2025201,
-      "arrivalLocation": "Main Entrance",
-      "age": 13,
-      "severity": "ESI3",
-      "patientName": "Michael",
-      "gender": "male",
-      "id": 194
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2042775,
-      "arrivalLocation": "Main Entrance",
-      "age": 40,
-      "severity": "ESI2",
-      "patientName": "Amy",
-      "gender": "female",
-      "id": 195
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2052000,
-      "arrivalLocation": "Main Entrance",
-      "age": 44,
-      "severity": "ESI3",
-      "patientName": "Melissa",
-      "gender": "female",
-      "id": 196
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2063007,
-      "arrivalLocation": "Main Entrance",
-      "age": 5,
-      "severity": "ESI3",
-      "patientName": "Liam",
-      "gender": "male",
-      "id": 197
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2090241,
-      "arrivalLocation": "Main Entrance",
-      "age": 6,
-      "severity": "ESI2",
-      "patientName": "Emma",
-      "gender": "female",
-      "id": 198
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2101654,
-      "arrivalLocation": "Main Entrance",
-      "age": 37,
-      "severity": "ESI2",
-      "patientName": "Christopher",
-      "gender": "male",
-      "id": 199
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2110772,
-      "arrivalLocation": "Main Entrance",
-      "age": 27,
-      "severity": "ESI3",
-      "patientName": "Andrew",
-      "gender": "male",
-      "id": 200
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2126419,
-      "arrivalLocation": "Main Entrance",
-      "age": 31,
-      "severity": "ESI3",
-      "patientName": "Jennifer",
-      "gender": "female",
-      "id": 201
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2135663,
-      "arrivalLocation": "Main Entrance",
-      "age": 60,
-      "severity": "ESI2",
-      "patientName": "Susan",
-      "gender": "female",
-      "id": 202
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2136102,
-      "arrivalLocation": "Main Entrance",
-      "age": 30,
-      "severity": "ESI3",
-      "patientName": "Michael",
-      "gender": "male",
-      "id": 203
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2138323,
-      "arrivalLocation": "Main Entrance",
-      "age": 12,
-      "severity": "ESI3",
-      "patientName": "Emily",
-      "gender": "female",
-      "id": 204
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2138600,
-      "arrivalLocation": "Main Entrance",
-      "age": 69,
-      "severity": "ESI3",
-      "patientName": "Mary",
-      "gender": "female",
-      "id": 205
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2159627,
-      "arrivalLocation": "Main Entrance",
-      "age": 20,
-      "severity": "ESI2",
-      "patientName": "Matthew",
-      "gender": "male",
-      "id": 206
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2160000,
-      "arrivalLocation": "Main Entrance",
-      "age": 17,
-      "severity": "ESI3",
-      "patientName": "Matthew",
-      "gender": "male",
-      "id": 207
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2172909,
-      "arrivalLocation": "Main Entrance",
-      "age": 23,
-      "severity": "ESI3",
-      "patientName": "Matthew",
-      "gender": "male",
-      "id": 208
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2176291,
-      "arrivalLocation": "Main Entrance",
-      "age": 61,
-      "severity": "ESI2",
-      "patientName": "Patricia",
-      "gender": "female",
-      "id": 209
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2179797,
-      "arrivalLocation": "Main Entrance",
-      "age": 19,
-      "severity": "ESI3",
-      "patientName": "Sarah",
-      "gender": "female",
-      "id": 210
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2182177,
-      "arrivalLocation": "Main Entrance",
-      "age": 51,
-      "severity": "ESI2",
-      "patientName": "Michael",
-      "gender": "male",
-      "id": 211
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2186980,
-      "arrivalLocation": "Main Entrance",
-      "age": 30,
-      "severity": "ESI3",
-      "patientName": "Brittany",
-      "gender": "female",
-      "id": 212
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2211286,
-      "arrivalLocation": "Main Entrance",
-      "age": 5,
-      "severity": "ESI3",
-      "patientName": "Jacob",
-      "gender": "male",
-      "id": 213
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2229873,
-      "arrivalLocation": "Main Entrance",
-      "age": 52,
-      "severity": "ESI3",
-      "patientName": "Lisa",
-      "gender": "female",
-      "id": 214
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2246522,
-      "arrivalLocation": "Ambulance Entrance",
-      "age": 9,
-      "severity": "ESI1",
-      "patientName": "Olivia",
-      "gender": "female",
-      "id": 215
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2249359,
-      "arrivalLocation": "Main Entrance",
-      "age": 54,
-      "severity": "ESI3",
-      "patientName": "Mary",
-      "gender": "female",
-      "id": 216
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2268000,
-      "arrivalLocation": "Main Entrance",
-      "age": 39,
-      "severity": "ESI2",
-      "patientName": "Jessica",
-      "gender": "female",
-      "id": 217
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2275397,
-      "arrivalLocation": "Main Entrance",
-      "age": 28,
-      "severity": "ESI2",
-      "patientName": "Samantha",
-      "gender": "female",
-      "id": 218
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2286841,
-      "arrivalLocation": "Main Entrance",
-      "age": 14,
-      "severity": "ESI3",
-      "patientName": "Joshua",
-      "gender": "male",
-      "id": 219
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2292966,
-      "arrivalLocation": "Main Entrance",
-      "age": 52,
-      "severity": "ESI3",
-      "patientName": "Michelle",
-      "gender": "female",
-      "id": 220
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2295147,
-      "arrivalLocation": "Main Entrance",
-      "age": 31,
-      "severity": "ESI3",
-      "patientName": "Ashley",
-      "gender": "female",
-      "id": 221
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2301724,
-      "arrivalLocation": "Main Entrance",
-      "age": 80,
-      "severity": "ESI3",
-      "patientName": "John",
-      "gender": "male",
-      "id": 222
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2322164,
-      "arrivalLocation": "Main Entrance",
-      "age": 65,
-      "severity": "ESI3",
-      "patientName": "Patricia",
-      "gender": "female",
-      "id": 223
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2322728,
-      "arrivalLocation": "Main Entrance",
-      "age": 51,
-      "severity": "ESI3",
-      "patientName": "Melissa",
-      "gender": "female",
-      "id": 224
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2331623,
-      "arrivalLocation": "Main Entrance",
-      "age": 32,
-      "severity": "ESI3",
-      "patientName": "Sarah",
-      "gender": "female",
-      "id": 225
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2346837,
-      "arrivalLocation": "Main Entrance",
-      "age": 19,
-      "severity": "ESI2",
-      "patientName": "Madison",
-      "gender": "female",
-      "id": 226
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2376000,
-      "arrivalLocation": "Main Entrance",
-      "age": 26,
-      "severity": "ESI3",
-      "patientName": "Christopher",
-      "gender": "male",
-      "id": 227
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2382157,
-      "arrivalLocation": "Main Entrance",
-      "age": 67,
-      "severity": "ESI3",
-      "patientName": "James",
-      "gender": "male",
-      "id": 228
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2389406,
-      "arrivalLocation": "Main Entrance",
-      "age": 8,
-      "severity": "ESI3",
-      "patientName": "Ava",
-      "gender": "female",
-      "id": 229
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2393101,
-      "arrivalLocation": "Main Entrance",
-      "age": 37,
-      "severity": "ESI3",
-      "patientName": "Christopher",
-      "gender": "male",
-      "id": 230
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2416187,
-      "arrivalLocation": "Main Entrance",
-      "age": 28,
-      "severity": "ESI2",
-      "patientName": "Amanda",
-      "gender": "female",
-      "id": 231
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2421623,
-      "arrivalLocation": "Main Entrance",
-      "age": 45,
-      "severity": "ESI2",
-      "patientName": "Jason",
-      "gender": "male",
-      "id": 232
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2422603,
-      "arrivalLocation": "Main Entrance",
-      "age": 54,
-      "severity": "ESI3",
-      "patientName": "Robert",
-      "gender": "male",
-      "id": 233
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2438432,
-      "arrivalLocation": "Main Entrance",
-      "age": 51,
-      "severity": "ESI2",
-      "patientName": "Melissa",
-      "gender": "female",
-      "id": 234
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2484000,
-      "arrivalLocation": "Main Entrance",
-      "age": 39,
-      "severity": "ESI3",
-      "patientName": "Amanda",
-      "gender": "female",
-      "id": 235
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2486298,
-      "arrivalLocation": "Main Entrance",
-      "age": 32,
-      "severity": "ESI3",
-      "patientName": "Amanda",
-      "gender": "female",
-      "id": 236
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2498216,
-      "arrivalLocation": "Main Entrance",
-      "age": 2,
-      "severity": "ESI3",
-      "patientName": "Ava",
-      "gender": "female",
-      "id": 237
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2510054,
-      "arrivalLocation": "Main Entrance",
-      "age": 28,
-      "severity": "ESI3",
-      "patientName": "Christopher",
-      "gender": "male",
-      "id": 238
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2510274,
-      "arrivalLocation": "Main Entrance",
-      "age": 82,
-      "severity": "ESI3",
-      "patientName": "James",
-      "gender": "male",
-      "id": 239
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2525020,
-      "arrivalLocation": "Main Entrance",
-      "age": 38,
-      "severity": "ESI2",
-      "patientName": "Melissa",
-      "gender": "female",
-      "id": 240
-    },
-    {
-      "name": "patient",
-      "arrivalTick": 2536970,
-      "arrivalLocation": "Main Entrance",
-      "age": 33,
-      "severity": "ESI2",
-      "patientName": "Ashley",
-      "gender": "female",
-      "id": 241
-    },
-    // EXTRA STAFF TO MEET DEMAND
-    // {
-    //   "name": "Tech",
-    //   "type": "Tech",
-    //   "arrivalLocation": "Main Entrance",
-    //   "arrivalTick": 10,
-    //   "id": 242
-    // },
-    // {
-    //   "name": "Tech",
-    //   "type": "Tech",
-    //   "arrivalLocation": "Main Entrance",
-    //   "arrivalTick": 1728000,
-    //   "id": 243
-    // },
-  ]
