+++ src/assets/index.js
@@ -1,16 +1,3 @@
-import arrivalHospital from "./arrivals/arrivalHospital.js"
-import arrivalHospitalFew from "./arrivals/arrivalHospitalFew.js"
-import arrivalsFlat from "./arrivals/arrivalsFlat.js"
-import arrivalHospitalTesting from "./arrivals/arrivalHospitalTesting.js"
-import arrivalHospitalThesis from "./arrivals/arrivalHospitalThesis.js"
-import arrivalHospitalThesisFinal from "./arrivals/arrivalHospitalThesisFinal.js"
-import arrivalHospitalThesisFinalCopy from "./arrivals/arrivalHospitalThesisFinalCopy.js"
-
-import locationsFlat from "./locations/locationsFlat.js"
-import locationsHospital from "./locations/locationsHospital.js"
-import locationsMinimal from "./locations/locations_minimal.js"
-import locationsDifferent from "./locations/1locations.js";
-
 import flatWall1 from "./objs/flat-wall-1.js"
 import flatWall2 from "./objs/flat-wall-2.js"
 import flat from "./objs/flat.js"
@@ -18,22 +5,18 @@ import hospital from "./objs/hospital.js"
 import locationsMinimalObs from "./objs/locations_minimal.js"
 import layoutDifferent from "./objs/1layout.js"
 
+import locationsDifferent from "./locations/1locations.js";
+import locationsFlat from "./locations/locationsFlat.js"
+import locationsHospital from "./locations/locationsHospital.js"
+import locationsMinimal from "./locations/locations_minimal.js"
+
+import arrivalHospital from "./arrivals/arrivalHospital.js"
+import arrivalHospitalFew from "./arrivals/arrivalHospitalFew.js"
+import arrivalHospitalFull from "./arrivals/arrivalHospitalFull.js"
+import arrivalHospitalTesting from "./arrivals/arrivalHospitalTesting.js"
+import arrivalsFlat from "./arrivals/arrivalsFlat.js"
+
 export default {
-  arrivals:{
-    arrivalHospital,
-    arrivalHospitalFew,
-    arrivalsFlat,
-    arrivalHospitalTesting,
-    arrivalHospitalThesis,
-    arrivalHospitalThesisFinal,
-    arrivalHospitalThesisFinalCopy
-  },
-  locations:{
-    locationsFlat,
-    locationsHospital,
-    locationsMinimal,
-    locationsDifferent
-  },
   objs:{
     flatWall1,
     flatWall2,
@@ -41,5 +24,18 @@ export default {
     hospital,
     locationsMinimalObs,
     layoutDifferent
+  },
+  locations:{
+    locationsFlat,
+    locationsHospital,
+    locationsMinimal,
+    locationsDifferent
+  },
+  arrivals:{
+    arrivalHospital,
+    arrivalHospitalFew,
+    arrivalsFlat,
+    arrivalHospitalTesting,
+    arrivalHospitalFull,
   }
 }
