+++ src/people/ct-thesis.js
@@ -3,7 +3,7 @@ import AssignComputer from "../behavior/assign-computer.js"
 import GoTo from "../behavior/go-to.js";
 import GoToLazy from "../behavior/go-to-lazy.js";
 import TakeTime from "../behavior/take-time.js";
-import task from "../support/task-thesis.js";
+import task from "../support/task.js";
 import fluentBehaviorTree from "@crowdedjs/fluent-behavior-tree"
 
 class ctThesis {
@@ -26,7 +26,12 @@ class ctThesis {
         let computer = Hospital.locations.find(l => l.name == goToName);
         let ct1 = Hospital.locations.find(l => l.name == "CT 1");
         let ct2 = Hospital.locations.find(l => l.name == "CT 2");
-        if (!myGoal) throw new exception("We couldn't find a location called " + goToName);
+        //if (!myGoal) throw new exception("We couldn't find a location called " + goToName);
+        if (!myGoal) { 
+            myGoal = ct1;
+            computer = ct1;
+        };
+
         let entrance = Hospital.getLocationByName("Main Entrance");
 
         let taskQueue = [];
