+++ src/people/greeter-nurse-thesis.js
@@ -6,7 +6,7 @@ import GoTo from "../behavior/go-to.js"
 import LookForArrivingPatient from "../behavior/look-for-arriving-patient.js";
 import TakeTime from "../behavior/take-time.js";
 import fluentBehaviorTree from "@crowdedjs/fluent-behavior-tree"
-import task from "../support/task-thesis.js";
+import task from "../support/task.js";
 
 class greeterNurseThesis {
