+++ src/people/nurse-thesis.js
@@ -3,7 +3,7 @@ import GoTo from "../behavior/go-to.js";
 import GoToLazy from "../behavior/go-to-lazy.js";
 import PatientTempState from "../support/patient-temp-state.js";
 import TakeTime from "../behavior/take-time.js";
-import task from "../support/task-thesis.js";
+import task from "../support/task.js";
 import fluentBehaviorTree from "@crowdedjs/fluent-behavior-tree"
 
 class nurseThesis {
