+++ src/people/old_trees/greeter-nurse-old.js
@@ -1,10 +1,10 @@
-import AssignPatientToTriageNurse from "../behavior/assign-patient-to-triage-nurse.js";
-import ComputerAssignPatientRoom from "../behavior/computer-assign-patient-room.js"
-import ComputerEnterPatient from "../behavior/computer-enter-patient.js";
-import ComputerScorePatient from "../behavior/computer-score-patient.js";
-import GoTo from "../behavior/go-to.js"
-import LookForArrivingPatient from "../behavior/look-for-arriving-patient.js";
-import TakeTime from "../behavior/take-time.js";
+import AssignPatientToTriageNurse from "../../behavior/assign-patient-to-triage-nurse-old.js";
+import ComputerAssignPatientRoom from "../../behavior/computer-assign-patient-room.js"
+import ComputerEnterPatient from "../../behavior/computer-enter-patient.js";
+import ComputerScorePatient from "../../behavior/computer-score-patient.js";
+import GoTo from "../../behavior/go-to.js"
+import LookForArrivingPatient from "../../behavior/look-for-arriving-patient.js";
+import TakeTime from "../../behavior/take-time.js";
 import fluentBehaviorTree from "@crowdedjs/fluent-behavior-tree"
