+++ src/people/old_trees/janitorial-old.js
@@ -1,8 +1,8 @@
-import GoTo from "../behavior/go-to.js"
-import GoToLazy from "../behavior/go-to-lazy.js";
+import GoTo from "../../behavior/go-to.js"
+import GoToLazy from "../../behavior/go-to-lazy.js";
 import fluentBehaviorTree from "@crowdedjs/fluent-behavior-tree"
-import LocationStatus from "../support/location-status.js";
-import TakeTime from "../behavior/take-time.js";
+import LocationStatus from "../../support/location-status.js";
+import TakeTime from "../../behavior/take-time.js";
 
 class janitorial {
