+++ src/people/old_trees/nurse-old.js
@@ -1,6 +1,6 @@
-import responsibility from "../behavior/responsibility/responsibility.js";
-import GoTo from "../behavior/go-to.js";
-import AssignComputer from "../behavior/assign-computer.js"
+import responsibility from "../../behavior/responsibility/responsibility.js";
+import GoTo from "../../behavior/go-to.js";
+import AssignComputer from "../../behavior/assign-computer.js"
 import fluentBehaviorTree from "@crowdedjs/fluent-behavior-tree"
 
 class nurse {
