+++ src/people/old_trees/phlebotomist-old.js
@@ -1,8 +1,8 @@
-import GoTo from "../behavior/go-to.js";
-import AssignBed from "../behavior/assign-bed.js";
-import AssignComputer from "../behavior/assign-computer.js";
-import responsibility from "../behavior/responsibility/responsibility.js";
-import WaitForever from "../behavior/wait-forever.js";
+import GoTo from "../../behavior/go-to.js";
+import AssignBed from "../../behavior/assign-bed.js";
+import AssignComputer from "../../behavior/assign-computer.js";
+import responsibility from "../../behavior/responsibility/responsibility.js";
+import WaitForever from "../../behavior/wait-forever.js";
 import fluentBehaviorTree from "@crowdedjs/fluent-behavior-tree"
