+++ src/people/old_trees/xray-old.js
@@ -1,23 +1,26 @@
-import GoTo from "../behavior/go-to.js"
-import WaitForever from "../behavior/wait-forever.js"
-import responsibility from "../behavior/responsibility/responsibility.js"
+import GoTo from "../../behavior/go-to.js"
+import AssignComputer from "../../behavior/assign-computer.js";
+import responsibility from "../../behavior/responsibility/responsibility.js";
 import fluentBehaviorTree from "@crowdedjs/fluent-behavior-tree"
 
-
-class radiology {
+class xray {
 
   constructor(myIndex) {
     this.index = myIndex;
-   
+
     const builder = new fluentBehaviorTree.BehaviorTreeBuilder();
-    this.toReturn = null;
 
     let self = this;//Since we need to reference this in anonymous functions, we need a reference
-    let goToName = "CT 2";
-    let me= ()=>Hospital.agents.find(a=>a.id == myIndex);
+    let goToName = "XRay 1";
+    if (myIndex % 2 == 1) {
+      goToName = "XRay 2";
+    }
 
+    let me= ()=>Hospital.agents.find(a=>a.id == myIndex);
     let myGoal = Hospital.locations.find(l => l.name == goToName);
-    if (!myGoal) throw new exception("We couldn't find a location called " + goToName);
+    if (!myGoal) throw new Exception("We couldn't find a location called " + goToName);
+
+    this.goTo = new GoTo(self.index, myGoal.location);
 
 
     this.tree = builder
@@ -30,33 +33,34 @@ class radiology {
           if (me().lengthOfStay == 43200 || me().lengthOfStay == 86399) {
             let idleTimeMinutes = ((1440 * me().idleTime) / 86400);
             idleTimeMinutes = Math.round((idleTimeMinutes + Number.EPSILON) * 100) / 100
-            //console.log("Radiology Idle Time: " + me().idleTime + " ticks / " + idleTimeMinutes + " minutes in-simulation");
+            //console.log("X-Ray Idle Time: " + me().idleTime + " ticks / " + idleTimeMinutes + " minutes in-simulation");
             console.log(idleTimeMinutes);
             me().idleTime = 0;
             //me().lengthOfStay = 0;
           }
           me().lengthOfStay++;
           return fluentBehaviorTree.BehaviorTreeStatus.Running; 
-      })
-    .sequence("Go and Idle")
+      })  
+    .sequence("X-Ray Tree")
       .selector("Check for arrival")  
         .condition("Clock in", async (t) => me().onTheClock)
         .do("SHIFT CHANGE", (t) => {
           // SHIFT CHANGE
           if (me().onTheClock == false) {
             me().onTheClock = true;
-            Hospital.activeRadio.push(me());
-            if (Hospital.activeRadio[0] != me() && Hospital.activeRadio.length > 1) {
-              for (let i = 0; i < Hospital.activeRadio.length; i++) {
-                if (!Hospital.activeRadio[i].replacement) {
-                  Hospital.activeRadio[i].replacement = true;
-                  //Hospital.activeRadio.shift();
-                  Hospital.activeRadio.splice(i, 1);
+            Hospital.activeXRay.push(me());
+            if (Hospital.activeXRay[0] != me() && Hospital.activeXRay.length > 2) {
+              for (let i = 0; i < Hospital.activeXRay.length; i++) {
+                if (!Hospital.activeXRay[i].replacement) {
+                  Hospital.activeXRay[i].replacement = true;
+                  //Hospital.activeXRay.shift();
+                  Hospital.activeXRay.splice(i, 1);
                   break;
                 }
               }
             }
           }
+          
           return fluentBehaviorTree.BehaviorTreeStatus.Success;
         })
       .end()
@@ -67,33 +71,35 @@ class radiology {
         .sequence("Exit Procedure")
           .splice(new GoTo(self.index, Hospital.locations.find(l => l.name == "Main Entrance").location).tree)
           .do("Leave Simulation", (t) => {
-            
-            // TESTING
-            let idleTimeMinutes = ((1440 * me().idleTime) / 86400);
-            idleTimeMinutes = Math.round((idleTimeMinutes + Number.EPSILON) * 100) / 100
-            console.log("Radiology Idle Time: " + me().idleTime + " ticks / " + idleTimeMinutes + " minutes in-simulation");
-            Hospital.radioData.push(me().idleTime);
-
+            // for(let i = 0; i < Hospital.computer.entries.length; i++) {
+            //   if (Hospital.computer.entries[i].getTech() == me()) {
+            //     Hospital.computer.entries[i].setTech(null);
+            //   }
+            // }
+            if (Hospital.aTeam[5] == me()) {
+              Hospital.aTeam[5] = null;
+            }
             me().inSimulation = false;
             return fluentBehaviorTree.BehaviorTreeStatus.Running;
           })
         .end()
       .end()
 
-      .splice(new GoTo(self.index, myGoal.location).tree)
-      .splice(new responsibility(myIndex).tree)
-      
+      .splice(this.goTo.tree)
+
+      .splice(new AssignComputer(myIndex, myGoal.location).tree) // XRay 1 or XRay 2
+
+      .splice(new responsibility(myIndex).tree) // lazy: true
+        
     .end()
     .end()
     .build();
   }
 
   async update( crowd, msec) {
-    //this.toReturn = null;//Set the default return value to null (don't change destination)
     await this.tree.tick({ crowd, msec }) //Call the behavior tree
-    //return this.toReturn; //Return what the behavior tree set the return value to
   }
 
 }
 
-export default radiology
+export default xray;
