+++ src/people/pharmacist-thesis.js
@@ -1,7 +1,7 @@
 import GoTo from "../behavior/go-to.js"
 import WaitForever from "../behavior/wait-forever.js"
 import fluentBehaviorTree from "@crowdedjs/fluent-behavior-tree"
-import task from "../support/task-thesis.js";
+import task from "../support/task.js";
 
 class pharmacistThesis {
 
@@ -14,11 +14,13 @@ class pharmacistThesis {
 
     let me= ()=>Hospital.agents.find(a=>a.id == myIndex);
     let myGoal = Hospital.locations.find(l => l.name == "Pharmacy");
-    if (!myGoal) throw new Exception("We couldn't find a location called Pharmacy");
+    //if (!myGoal) throw new Exception("We couldn't find a location called Pharmacy");
     let entrance = Hospital.getLocationByName("Main Entrance");
+    if (!myGoal) {
+      myGoal = Hospital.locations.find(l => l.name == "B Desk");
+    }
 
     this.goTo = new GoTo(self.index, myGoal.location);
-
 
     this.tree = builder
       .sequence("Pharmacist Tree")
