+++ src/simulations.js
@@ -5,8 +5,7 @@ export default [
     //"arrivalPath" : "arrivalHospital.json",
     //"arrivalPath" : "arrivalHospitalTesting.json",
     //"arrivalPath" : "arrivalSecondPatient.json",
-    //"arrivalPath" : "arrivalHospitalThesis.json",
-    "arrivalPath" : "arrivalHospitalThesisFinal.json",
+    "arrivalPath" : "arrivalHospitalFull.json",
     "secondsOfSimulation" : 4500,
     "locationsPath" : "locationsHospital.json"
   },
@@ -14,7 +13,21 @@ export default [
     "title":"Flat Test",
     "objPath" : "flat.obj",
     "arrivalPath" : "arrivalsFlat.json",
-    "secondsOfSimulation" : 2500,
+    "secondsOfSimulation" : 4500,
     "locationsPath" : "locationsFlat.json"
+  },
+  {
+    "title":"Altered Layout Test",
+    "objPath" : "1layout.obj",
+    "arrivalPath" : "arrivalHospitalFull.json",
+    "secondsOfSimulation" : 4500,
+    "locationsPath" : "1locations.json"
+  },
+  {
+    "title":"Minimal Layout Test",
+    "objPath" : "locations_minimal_floorplan.obj",
+    "arrivalPath" : "arrivalHospitalFull.json",
+    "secondsOfSimulation" : 4500,
+    "locationsPath" : "locations_minimal.json"
   }
 ]
