+++ src/support/hospital.js
@@ -1,5 +1,5 @@
 import * as THREE from "three"
-import task from "./task-thesis.js";
+import task from "./task.js";
 
 class Hospital{
