+++ task_flowchart/HospitalTasksFlowchart.drawio
@@ -1,0 +1,1 @@
+<mxfile host="app.diagrams.net" modified="2022-02-17T02:13:28.571Z" agent="5.0 (Windows)" etag="MGyIoM7EytoZ51ImBcKp" version="16.5.6" type="device"><diagram id="hZtonnguE62RwNzXROfG" name="Page-1">7V1Zd6q+1/40rvW+F+0ijHJpHTpPtna6OStiVFoUC1htL36f/Z8wKKDitkcUPPTiHIlM7jz72UOSnZJQHUxPLTzqX5sdYpR4rjMtCbUSz4ucKtD/WMu314JkBXktPUvv+G3zhgf9h/iNnN861jvEjpzomKbh6KNoo2YOh0RzIm3YssxJ9LSuaUSfOsI9/4ncvOFBwwZZOO1Z7zh9r7Ushc4+I3qvHzwZcf43Axyc7DfYfdwxJ6EmoV4SqpZpOt6nwbRKDCa9QC7edY0V385ezCJDB3KB9HR58XFdJuTP+YVsvTv9+5Z9hGT/Pl/YGPs/2X9d5zuQgWWOhx3CbsOVhJNJX3fIwwhr7NsJ7Xba1ncGBj1C9GNXN4yqaZgWPR6aQ3rSif8EYjlkuvLl0UwkFEzEHBDH+qanBBegQOI+kHhR8o4n815Bgn9OP9wjQSP2kdCb3XwuLPrBl9dGsuNyITsFSb8TnaCkJrr1gutRyTHxdCxz9IitHnF8GY5MfegQq/5Ff77tt21BSoIYBZgkLAhJFpcIKRDmBjJq2cS6bb8zuuI5A7eJ4V16Ztoj3cFMeo/Y/rDp/w3DnGh9bDneQ4yxpnfcY+/6c1/ID2N8wg1uRq9IuPTOTBQ/H5KE310hyYcAaVKRdg2XtdqGqX3QJvd/BmF92PPPWobqrjl0fDZH6rHitwTgLvECVnG3XQ7O9J/O7oYNvTeknzXCupk22N7DrkiXiVyZtzT9Xgg1PZr04TV+3kCPBHbkWOYHuWVtzrcPGqZt0Za5srov0rZNY+yQiqUFv4S1zo/4mMrSX9V1/2YPDEwGOpagKF2tL4vQXYdNAW2MTXoYgmciXEu8pFkEO6Z1gweEHpWYBeJKStX/wEsd7JCScjKgf8wJKPHsq2/6V1Jq7gXurzcocvrUMBMraqyT0P7z2eiPMDIBaBd2i3avr1ehvUA3GN3fURCvAzsS0wX7OkS2hpMLh1yKAESKUES+E8f5DjqkDACf4B838EA3mADPiPFFHF3Ds15bhso4gJgXu7yXo/6M7VBRVJibPfdYyLATa/EkQjqBp7091pt52xEgBI0WMbCjf0Wfugwe7jPoS+Pv0Amun2FHXeo71hbySlUU8RnirqZtji2N+Bcl3EeUxMT7OK7/s3AfF80zwfwlwO+oqIj7UnXGTHYy4I/uccvqX/Ij6UwFAF5Kn4LpXWlMSFboQwj03l8CFTNSPS4v05fZpcv1JUq5Xtvfk64mdBWtvKCOnPuKW3B8j+KRlboYHshL1EzePDjYGt86LdyqW0/fP7x8D4CfnCrfbgqTZf0Yg8kMELzIoE2ZM/CQTcvpmz1ziI36vHUdDweqRJsauhH8WDLVnReX2DlO9o9f3TeS/KM7Yum0T5lyuHRPEWV9s0uOOHqNELSELmKH8asA9O83eXy5jkd4zgniwVXnqRmzHL9ialgorZQWQ2nZcFjkrH9FgC5/jk2v3aYuN32NCkuIGYw4aB9x72Pb0bvfRxqFt9sx7OsZNzKpk6lz5JNm/Du3+YjqysD2vmxjmxi6b/45pjJHtgtp9qWHCo59fxTIn7WjY56haEC7Vx8eOczbZc1H/Mh7x+AX0E89/3/3l1LiHi5vW/Lzw++CBPfOmqe7lbDuLj7tGRsfTAw05qCEIkwroafR7lv2Equb3c6Jthb9td3+qvaJxjrsnN23SgmW/fqt9FcsVcUEXYok8ixCXxi33ROY8FzDdoK1j55L9IGt6JAuHrtPWPBFmEmnXrtR8b8Y6J2OAU4IrrWCS9wA7hgJfFkQyhLiJFFSyrywIRlumO46tYhrJXjuZmxRzy3R4n8/41bfak4m170pwOKXdxvzJ7mZ0PwWbVKXeRN8mdd27nSuiPS35nTO0qaB0xk4oWETzS9LaaXqdV6Z5odLfwyV1KzrX6xjeG4WHCVB1KIQrb+Rifn1PABAVM0ORDeF5F7ioLQhiWQ1CskyEJJqqpCsmoPR2KPJumfBYGh8esGtqnNn3PJKDYBGxBVwzDQcRS4TcEwG3QsF3ZOk3OpPzxIEdAgKOlBgDkPmoYXv83CdxeEsmBfXhe+hYJ1d40bz4g7CdxUYvkfIKLfxO4DRHzTTIlBGH73hlm2q7cpDswdRrh2P8haMviGjS3wOGP2Lgs55uZ/QEwQI6MCDrQWjHwKjB9P01lM6v1w3Do7SK7bNuC7E6VzTNAfJWnbOtKzRudZ+5BuIloEHkAtq3wu1zyYoZJnarynozp6r9vddEzJpAYEHcQtqPwhq56HULi7XjcOg9iV8zkaAuEdLZ1PHQclrjHBLvnE0dM6/QTQNPF5d0LtP750OltOjdwFlMjWYDLpbCrq36at2OuBBCUBwPjpD9N7Bdn+GnK4+rdFj/4h9RVWWgpbBl6kUtwoh+558IYRMAmDyRcwgoF3ZAxFqD1YMNB6GPdiE91943OLvulxH+k8BqCCfoRx8MWq5xA7MMjbZGrV8xjpr8UYt526K57kM1uYVRYrSSfXk/OvP6wUEpeCkfeGd7AeVefBO/hOod/KnVn3RLzkI6Iq84l58jX1N9PR8CICzsUo3DsTZwB+R4SEv9FyfTzwRcat52Wg1hqYN0a4in5htSs/F4P81Bd3tS6f5fYFOIaAr8ol7yieKqlLaR0ZxVsBhLakfdEbxiuAv8KD/C1WqR2Kj6rUCmVHDF6nDbKUO40yei0F/joLupy4+/Ag6ZCYrrxRM/i+NDPHQTCB/0JlAQAKwKVGXSLo7u8MyiL2LZQvZ9suVbCYATwlrOCPYoC/BFjx1TWtAFcccroenWL04bb6+gDz2YslCxuGYjcxfTbcZ2HpgJ5fhUDDfL87eLyHpCqHIQGfLyY3jMBfpCga688HT/XVzVIOAjs+hkwseHxfor1lVYyJdrzhoyWVCA7qgQTjoLHXd1kxWoS6ep65PdQDxm/zDH/H92oHo4I5LeeWe+KtCQ6mmWDsmTvy5yG4w0JWV28eLqzOQt7Hdal1FdiPj2Q0BmqWOsNHBkfoj0frJeoSYHjUtRG5eryF6tIMiYAed3JBrJ6jRSJHMY7NcFSFDyY0nVhh3TVE6BsjmxVB8VwXIdDuhGCvJWDojDkB5p94ErN6W7C+DDCFlve2MVs2cn3NluhVd484DHlP/OVJQaDaFJ2RF5za1Ng2MJzsI6r3OLWnYjobM6vwy9yi4bpu2tAy0pdKKEd9tmUj/zrG6m2U5WoVG9UsSr6rTGT9fiJ5PP3hvALTGyYxXvzxdT3f48/n4h/uBrCoRMjS4kAu6azRqtUqaWbMY3ZWzkb1dDzrtZ3jBNWoGBHQHvaqkyJr9yihA53ZGGOvgAqw73a1O2BpFEmf/d0K6Xk0QL6vm1R977JPB/69XzLfPyduHKR0BFFPM0FqTXFiD1MdQYtaAvndOzMFHrzx6UwZjCOryWPDpHzEHe8q2QSOECGEdnDFI1rKKjFvOtaOiXvkTomV5HKrMWsY6R4tlePBimYOoi/47HTqiOnR2fzm6a943ITpULDgrdGht1vjgdIjtO0dbrnSbPYT+ADcUIVPNGHfYp6rb335NdfYK86bbMbvXCenjL91k+8dwdxaxvajGPb2iOWNvczuLkDWp/H6Zauv1iW2flCFpBnEHC9hm+3cu09Qt7THjK9hCIqperaiL+r2dwEMoR5OaSzY+XBp2zBr3EXd0Vdy6vjk7eb+/eoHgo1hrVrD5v5diqj7Sk1+a7iNZumk8StaqN9xuHX9OPz4rMmTdj1gMoGY8h4R2Ox8LuPmvlDyESoy2OSlGT6l+KUAWU/jlsEh39FSN7eGBeOGvhkOh6BFzgJ59oEBdMR9txygQ/g4F8EnGSZZMppZMn0z/VE5fEcSSgVewFpbMt2S7XdmE+GxU1K1PHQsPiTl2N2cfDzW2yI59NnS3Ussj/vBH7HS2JTbb7vrUrRZqhs732vGwQ/+9H5Oxd6G34TsOB9zshDM3BE/C+n8U63dPb8/3R89PEKzvYB7IocbKMxTmKVa+0dqt+mtzeHzzXIfgI49TNrIWKwcOQcxnXOVPJnuhO/A0oWX8I4by4OLlZD1qUT2a9O96bfG8BdAjCTzDIkN69K+MdUeTUjsb7Ybr2UGPMiTrmUT1zNDF48l/+AqiZ3mcU5I5e/W7CYRoP1oEze4etrVqElvvrI2E3zrt1sundWHgJ4j3J2VoiW0uV2l1u0QQ+fQiY3W/JWiAi2T8PHRkV/qEpF1Kq2a2yDlB1nH9SpYVljutXbTdZSJck+DOeh64bZ/3X5sfMoQHipKw2VotsqD32Vgt4sGvag5ttpH9WgSeSuTm/LVdhiCwGF3MNgIzUuXHQ+Ct1XG33/NHu/2pSZXH9ZD8kXp3D8NPyMRaqVhCl61hgjgkc1F/hIFOr/O1r+NmBQK6Ih/7T9UfkaDRrXQQc5eAo/5B9HI4EYUMzQXKKzgwrYhinlfgHjQ8BEUXr4QaUn76eGG03gGcJherD7Pt2+14A1soCaCDIwFosSmZ3y0JVByHWlJvLgMoujumDDC2MDk7szCEAYoartlmgEwW30HCXhQ+YVR//fzRyBjJTuePyjyQXFZVsku5+k55w+o78YUqcioTDee1GFLhvTCPF7yXPd7LSBWeKiBztcGwnpIhhzuXw3ppF19EiIvNo0ZIggFxlwN7KFhZlaPyd/szwArUAK+a7pOuAaadGcecT2qrTPDiFWI52WircVhvfAH6OzMPxbW6V8cymp6E4Tq+D3wGcY3QXuo6zjAzXz4gr0FdmUu8YFvOpTsixdXMWW5tnY3fYMg+sgqscC3371ouseki0Kan61x6g6RbdC/VDGVzCvcSBEUhm+6lUriXYDOsChl3L/k45tRkK7zoXm5+xcE4pOXCId2+JmTGIRWlDR3S2AXbckj9+VLpeKRqhiaTFx7pcjcgGx7pBR7qDpXCNp1SxBVe6d/BUVUbjXS9UikKx4zsp/vA0MjEsk0yRFzBhpuy4a7hl42xnzuKtQEVps3avTIR4ekYfvlEv1TE0GRn6cNZwYj+tgtGILQDHj3UihFq1I3LSMGIJu7oVAK97WaBEMpQBjKXBjftNFB8sJvapAyuHkT8XvfY+tUsn/1VifNDWsgcwhUl9FMuELaQP+E2S7jI5VQi3xAHck3ypZNJGiFwNOFQeH37j4EXOXC3Mx1/t2joFuHW2/RVOx3wNQjs4Lu9ZWjVELj6DKKqx63CyF5rJC/ajizXfZKgu6+sqjC687VGK/KrXCy/Kq0Z71+4ID71JWplFq7nVfVYkNT5X9Ri+aMWs/f3OsK/R8JrCeXojYQydxxslxbcy+urhXvtsEDPxlyUxxWMBRdllYtW1bk9HC4SNuIigUviIoX7LRfFbkRfKoNc1JRw63zwdH/dHIG4CL7vYR65KKtV+SLcEx0l5nLFPatC6L2tuV5QWyWWcVzmQkApQCmrx7LIzf6EyJ1Fjn2bNULY1DmBl5nJIyFk1jmBEwRg8skONrIHTzDhsu2dKLF5JfKaRXeIR1LSBWt8E8RJsQEQ5ZfuyHw9zGxGDBfLx2yTfKCbUfiGI2lFejQxvYw0QiQTTob5OlXaQs4rJjt+2awPYQkqlV/k/aGiA9QIzIDolCC9HvjTPFB0MkpNdMHC3oyLLo46xEFlh9KDnSzkQnZoZg/m6wGXpKl3LTwxp8ITOCjyxPSEJ+VCeNKCrQDibraRZAqik3MhuoX5REvM7GwUKSw6Ke7KwIaWLNN0wj4O/Zn9a7PDAqX6/wA=</diagram></mxfile>
